ONSAS

Documentation for ONSAS.jl.

ONSAS.Geometry

ONSAS.Material

ONSAS.ONSAS_init

ONSAS.ONSAS_solve

ONSAS.assembler

ONSAS.computeElemsByMEBI

ONSAS.displacements

ONSAS.linear_truss

ONSAS.nodes2dofs

ONSAS.Geometry — Type

Struct with information of the geometry of the element:

element_type: a string with: node, truss, frame,triangleortetrahedron`.

ONSAS.Material — Type

Struct with information of the material.

ONSAS.ONSAS_init — Method

Function used to initialize the model solution and properties structs.

ONSAS.ONSAS_solve — Method

Function used to solve the problem and generate a vector of ModelSolution

ONSAS.assembler — Method

assembler is the function in charge of the construction of the tangent matrix and the RHS vector of the main FEM iteration system

The inputs are:

nodalCoords: a matrix that in the i-th colum contains the x, y and z coordinates as a column vector, for the i-th node of the mesh.

elemNodalConnec: a vector, that in the i-th entry contains a vector of the indexes of the node-connectivity of that element.

MEBIValsMat: a matrix of indexes, at each column contains the the four MEBI parameters

ONSAS.computeElemsByMEBI — Method

computeElemsByMEBI This function computes the vector of the element indexes for a given vector of MEBI parameter

ONSAS.displacements — Method

displacements(sol::ModelSolution)

Return the vector of displacements of the given solution.

ONSAS.linear_truss — Method

function to linear truss element

ONSAS.nodes2dofs — Method

nodes2dofs computes the vector of dofs for an input vector of nodes