Boundary conditions

ONSAS.BoundaryConditions — Module

Module defining the boundary conditions implemented. Two types of boundary conditions are defined Load (Neumann), and Displacements (Dirichlet). Overall, each boundary condition consists of a data type with a label, dofs and values into its fields.

ONSAS.BoundaryConditions.AbstractBoundaryCondition — Type

Abstract supertype for all elements.

An AbstractBoundaryCondition object facilitates the process of defining:

- Dirichlet or displacements boundary conditions.
- Neumann or load boundary conditions.

Abstract Methods

apply

Base.values

dofs

label

ONSAS.BoundaryConditions.AbstractDirichletBoundaryCondition — Type

Abstract supertype for all Dirichlet boundary conditions.

ONSAS.BoundaryConditions.AbstractNeumannBoundaryCondition — Type

Abstract supertype for all Neumann boundary conditions.

ONSAS.BoundaryConditions.AbstractNeumannBoundaryCondition — Method

Abstract functor for a AbstractNeumannBoundaryCondition that evaluates the load at time t.

Base.values — Method

Return the values function imposed to the respective degrees of freedom. This should be a function of time returning a vector with the same size as the Node or Element Dofs.

ONSAS.BoundaryConditions.apply — Method

Apply the boundary condition bc to anentity.

ONSAS.Utils.dofs — Method

Return the degrees of freedom symbol where the boundary condition is imposed

ONSAS.Utils.label — Method

Return the boundary condition label