Local loads

ONSAS.LocalLoadBoundaryConditions — Module

Module to handle loads in local coordinates.

ONSAS.LocalLoadBoundaryConditions.Pressure — Type

Pressure boundary condition imposed in local coordinates to the AbstractElement or AbstractFace. This is a force along the minus normal direction of the face.

ONSAS.BoundaryConditions.apply — Method

Return the dofs and the values imposed in the GlobalLoadBoundaryCondition lbc to the AbstractFace f at time t.