Cross-sections Interface

This section describes the interfaces for different cross-section types.

ONSAS.CrossSections — Module

Module defining geometric entities interface. Each cross section consists of a data type with one or more geometric parameters into its fields.

ONSAS.CrossSections.Circle — Type

Circle cross-section.

Fields:

d – circle diameter.

ONSAS.CrossSections.GenericCrossSection — Type

Generic cross-section.

This generic cross-section struct can be used to define a cross-section not belonging to the existing categories.

Fields:

area – area.

Ixx – moment of inertia respect to x axis.

Iyy – moment of inertia respect to y axis.

Izz – moment of inertia respect to z axis.

Ixy – product moment of area respect to the x-y axes.

Ixz – product moment of area respect to the x-z axes.

Iyz – product moment of area respect to the y-z axes.

ONSAS.CrossSections.GenericCrossSection — Method

Constructor of a GenericCrossSection if x, y and z are the principal axes.

ONSAS.CrossSections.Rectangle — Type

Rectangle cross-section.

Fields:

width_y – width in y local axis.

width_z – width in z local axis.

ONSAS.CrossSections.Square — Type

Square cross-section.

Fields:

width – width in y and z axes.

ONSAS.CrossSections.I_tensor_xyz — Method

Return the inertia tensor in the system of coordinates x-y-z

ONSAS.CrossSections.Ixx — Method

Return the moment of inertia of the AbstractCrossSection cs with respect to the local x axis.

ONSAS.CrossSections.Ixx — Method

Return the moment of inertia of a Circle cross-section c with respect to the local x axis.

ONSAS.CrossSections.Ixx — Method

Return the moment of inertia of a Rectangle cross-section r with respect to the local x axis.

ONSAS.CrossSections.Ixx — Method

Return the moment of inertia of a Square cross-section s with respect to the local x axis.

ONSAS.CrossSections.Ixy — Method

Return the product moment of area of the AbstractCrossSection cs respect to the local x-y axes.

ONSAS.CrossSections.Ixy — Method

Return the product moment of area of a Circle cross-section c with respect to the local x and y axes.

ONSAS.CrossSections.Ixy — Method

Return the product moment of area of a Rectangle cross-section r with respect to the local x and y axes.

ONSAS.CrossSections.Ixy — Method

Return the product moment of area of a Square cross-section s with respect to the local x and y axes.

ONSAS.CrossSections.Ixz — Method

Return the product moment of area of the AbstractCrossSection cs respect to the x-z local axes.

ONSAS.CrossSections.Ixz — Method

Return the product moment of area of a Circle cross-section c with respect to the local x and z axes.

ONSAS.CrossSections.Ixz — Method

Return the product moment of area of a Rectangle cross-section r with respect to the local x and z axes.

ONSAS.CrossSections.Ixz — Method

Return the product moment of area of a Square cross-section s with respect to the local x and z axes.

ONSAS.CrossSections.Iyy — Method

Return the moment of inertia of the AbstractCrossSection cs with respect to the local y axis.

ONSAS.CrossSections.Iyy — Method

Return the moment of inertia of a Circle cross-section c with respect to the local y axis.

ONSAS.CrossSections.Iyy — Method

Return the moment of inertia of a Rectangle cross-section r with respect to the local y axis.

ONSAS.CrossSections.Iyy — Method

Return the moment of inertia of a Square cross-section s with respect to the local y axis.

ONSAS.CrossSections.Iyz — Method

Return the product moment of area of the AbstractCrossSection cs respect to the y-z axes.

ONSAS.CrossSections.Iyz — Method

Return the product moment of area of a Circle cross-section c with respect to the local y and z axes.

ONSAS.CrossSections.Iyz — Method

Return the product moment of area of a Rectangle cross-section r with respect to the local y and z axes.

ONSAS.CrossSections.Iyz — Method

Return the product moment of area of a Square cross-section s with respect to the local y and z axes.

ONSAS.CrossSections.Izz — Method

Return the moment of inertia of the AbstractCrossSection cs respect to the local z axis.

ONSAS.CrossSections.Izz — Method

Return the moment of inertia of a Circle cross-section c with respect to the local z axis.

ONSAS.CrossSections.Izz — Method

Return the moment of inertia of a Rectangle cross-section r with respect to the local z axis.

ONSAS.CrossSections.Izz — Method

Return the moment of inertia of a Square cross-section s with respect to the local z axis.

ONSAS.CrossSections.area — Method

Return the cross-section area of the AbstractCrossSection cs

ONSAS.CrossSections.area — Method

Return the area of a Circle cross-section c.

ONSAS.CrossSections.area — Method

Return the area of a Rectangle cross-section r.