Circles

ONSAS.Circles.Circle — Type

Circle cross-section.

ONSAS.CrossSections.Ixx — Method

Return the moment of inertia of a Circle cross-section c with respect to the local x axis.

ONSAS.CrossSections.Ixy — Method

Return the product moment of area of a Circle cross-section c with respect to the local x and y axes.

ONSAS.CrossSections.Ixz — Method

Return the product moment of area of a Circle cross-section c with respect to the local x and z axes.

ONSAS.CrossSections.Iyy — Method

Return the moment of inertia of a Circle cross-section c with respect to the local y axis.

ONSAS.CrossSections.Iyz — Method

Return the product moment of area of a Circle cross-section c with respect to the local y and z axes.

ONSAS.CrossSections.Izz — Method

Return the moment of inertia of a Circle cross-section c with respect to the local z axis.

ONSAS.CrossSections.area — Method

Return the area of a Circle cross-section c.