Cross sections

ONSAS.CrossSections — Module

Module defining geometric entities interface. Each cross section consists of a data type with one or more geometric parameters into its fields.

ONSAS.CrossSections.AbstractCrossSection — Type

Abstract supertype for all elements corss-section.

Common methods

The following functions must be implemented for a new cross-section:

area – Return the cross-section area.

Ixx – Return the moment of inertia with respect to x axis.

Iyy – Return the moment of inertia with respect to y axis.

Izz – Return the moment of inertia with respect to z axis.

Ixy – Return the product moment of area respect to the x-y axes.

Ixz – Return the product moment of area respect to the x-z axes.

Iyz – Return the product moment of area respect to the y-z axes.

I_tensor_xyz – Return the inertia tensor in the system of coordinates x-y-z.

ONSAS.CrossSections.I_tensor_xyz — Method

Return the inertia tensor in the system of coordinates x-y-z

ONSAS.CrossSections.Ixx — Method

Return the moment of inertia of the AbstractCrossSection cs with respect to the local x axis.

ONSAS.CrossSections.Ixy — Method

Return the product moment of area of the AbstractCrossSection cs respect to the local x-y axes.

ONSAS.CrossSections.Ixz — Method

Return the product moment of area of the AbstractCrossSection cs respect to the x-z local axes.

ONSAS.CrossSections.Iyy — Method

Return the moment of inertia of the AbstractCrossSection cs with respect to the local y axis.

ONSAS.CrossSections.Iyz — Method

Return the product moment of area of the AbstractCrossSection cs respect to the y-z axes.

ONSAS.CrossSections.Izz — Method

Return the moment of inertia of the AbstractCrossSection cs respect to the local z axis.

ONSAS.CrossSections.area — Method

Return the cross-section area of the AbstractCrossSection cs