Generic cross sections

ONSAS.GenericCrossSections.GenericCrossSection — Type

This generic cross-section struct can be used to define a cross-section not belonging to the existing categories.

ONSAS.GenericCrossSections.GenericCrossSection — Method

Constructor of a GenericCrossSection if x, y and z are the principal axes.