Elements Interface

This section describes the interfaces for different element types.

ONSAS.Elements — Module

Module defining element formulations implemented. Each element consists of a data type with the geometric properties such as nodes, cross-sections and a label into its fields.

ONSAS.Elements.MINOR_INDICES — Constant

Indices for computing the minors of the interpolation matrix, implemented as a hash table.

ONSAS.Elements.AbstractElement — Type

Abstract supertype for all elements. An AbstractElement object facilitates the process of evaluating: - The internal forces vector and its tangent matrices. - The inertial forces vector and its tangent matrices. - The mechanical stresses and strains.

Common methods:

coordinates

dimension

dofs

local_dofs

These methods is a hard contract and must be implemented to define a new element.

local_dof_symbol

label

nodes

weights

Base.:∈

This method is a hard contract and for static analysis must be implemented to define a new element.

internal_forces

This method is a hard contract and for dynamic analysis must be implemented to define a new element.

inertial_forces

Common fields:

nodes

label

ONSAS.Elements.AbstractEntity — Type

Abstract supertype for all Faces and Elements.

An AbstractEntity object is an entity defined by dofs and node/s with certain coordinates and dimension.

Common methods:

create_entity

coordinates

dimension

dofs

Common fields:

nodes

label

ONSAS.Elements.AbstractFace — Type

Abstract supertype for all elements.

An AbstractFace object facilitates the process of adding boundary conditions on a surface.

Common methods:

area

create_entity

coordinates

dofs

label

nodes

normal_direction

Common fields:

nodes

label

ONSAS.Elements.AbstractNode — Type

An AbstractNode object is a point in space.

Common methods:

coordinates

dimension

dofs

ONSAS.Elements.Dof — Type

Scalar degree of freedom of the structure.

ONSAS.Elements.Node — Type

A Node is a point in space. The coordinates of the node are stored using a static array.

ONSAS.Elements.Node — Method

Node constructor with an AbstractVector v.

ONSAS.Elements.Node — Method

Node constructor with a NTuple t and an optional Dof dictionary dofs.

ONSAS.Elements.Node — Method

1D Node constructor with coordinates x₁.

ONSAS.Elements.Node — Method

3D Node constructor with coordinates x₁, x₂ and x₃ and an optional Dof dictionary dofs..

ONSAS.Elements.Node — Method

2D Node constructor with coordinates x₁ and x₂.

ONSAS.Elements.Point — Type

Construct a point given its coordinates.

It is a type alias for a statically sized array of dimension dim and element type T. Use either as a vararg function, Point(1, 2, 3), or by splatting Point(x...) if x is an (abstract) vector.

ONSAS.Elements.Tetrahedron — Type

Constructor for a Tetrahedron element without nodes and a label. This function is used to create meshes via GMSH.

ONSAS.Elements.Tetrahedron — Type

A Tetrahedron represents a 3D volume element with four nodes.

See [Belytschko] and [Gurtin] for more details.

ONSAS.Elements.TriangularFace — Type

Constructor for a TriangularFace element without nodes and a label. This function is used to create meshes via GMSH.

ONSAS.Elements.TriangularFace — Type

A TriangularFace represents an element composed by three Nodes.

ONSAS.Elements.TriangularFace — Method

Constructor for a TriangularFace element considering the nodes n₁ n₂ and n₃.

ONSAS.Elements.Truss — Type

Constructor for a Truss element without nodes and a label. This function is used to create meshes via GMSH.

ONSAS.Elements.Truss — Type

A Truss represents an element composed by two Nodes that transmits axial force only.

References

See [ANLE].

ONSAS.Elements.Truss — Method

Constructor for a Truss element considering the nodes n₁ and n₂ and the cross-section g.

Base.:∈ — Method

Return true if a point p is inside the AbstractElement e.

Base.:∈ — Method

Checks if a point p is inside a Tetrahedron element t.

Base.getindex — Method

Return the coordinate at component i from the AbstractNode n.

Base.size — Method

Return a tuple containing the dimensions of the AbstractNode n.

Dictionaries.index — Method

Return the dof index of the Dof d

ONSAS.CrossSections.area — Method

Return the AbstractFace f area.

ONSAS.CrossSections.area — Method

Return the area of a TriangularFace element tf.

ONSAS.Elements._X_rows — Method

Return auxiliar matrices with the element coordinates at the deformed and reference configurations.

ONSAS.Elements._area_vec — Method

Return the area vector with direction and modulus of a TriangularFace element tf.

ONSAS.Elements._aux_b — Method

Return auxiliar vectors bref and bdef of a truss element.

ONSAS.Elements._aux_matrices — Method

Return

ONSAS.Elements._coordinates_matrix — Method

Return the reshaped coordinates elem_coords of the tetrahedron element into a 4x3 matrix.

ONSAS.Elements._jacobian_mat — Method

Computes Jacobian matrix

ONSAS.Elements._lengths — Method

Return deformed and reference lengths of a truss element.

ONSAS.Elements._shape_functions_derivatives — Function

Return the shape functions derivatives of a Tetrahedron element.

ONSAS.Elements._strain — Method

Return the rotated engineering strain for a given reference and deformed length l_ini and the deformed length l_def.

ONSAS.Elements._volume — Method

Computes volume element of a tetrahedron given J = det(𝔽).

ONSAS.Elements.coordinates — Method

Return the AbstractEntity e coordinates.

ONSAS.Elements.coordinates — Method

Return the AbstractNode n coordinates.

ONSAS.Elements.coordinates — Method

Return each AbstractNode coordinates in a Vector of Nodes vn.

ONSAS.Elements.coordinates — Method

Return each AbstractFace coordinates in a Vector of Faces vf.

ONSAS.Elements.create_entity — Method

Return an AbstractEntity given an empty AbstractEntity e and a Vector of Nodes vn.

ONSAS.Elements.create_entity — Method

Return a Tetrahedron given an empty Tetrahedron t and a Vector of Nodes vn.

ONSAS.Elements.create_entity — Method

Return a TriangularFace given an empty TriangularFace tf and a Vector of Nodes vn.

ONSAS.Elements.create_entity — Method

Return a Tetrahedron given an empty Tetrahedron t and a Vector of Nodes vn.

ONSAS.Elements.cross_section — Method

Return the AbstractElement e cross_section.

ONSAS.Elements.cross_section — Method

Return the cross-section of a Truss element t.

ONSAS.Elements.dimension — Method

Return the AbstractEntity dimension.

ONSAS.Elements.dimension — Method

Return the AbstractNode n dimension (1D, 2D or 3D).

ONSAS.Elements.inertial_forces — Method

Return the inertial forces vector of an AbstractElement e. with an AbstractMaterial m

ONSAS.Elements.internal_forces — Method

Return the internal force of a Tetrahedron element t doted with an AbstractHyperElasticMaterial m + and a an element displacement vector u_e.

ONSAS.Elements.internal_forces — Method

Return the internal forces vector of an AbstractElement e with an AbstractMaterial m.

ONSAS.Elements.internal_forces — Method

Return the internal force of a Tetrahedron element t doted with an LinearIsotropicMaterial m.

Arguments

material: IsotropicLinearElastic type, the linear elastic material of the tetrahedron element.

element: Tetrahedron type, the tetrahedron element for which internal forces are to be computed.

displacements: AbstractVector type, the nodal displacements of the element.

Return

A 4-tuple containing:

forces: Symmetric type, the internal forces of the tetrahedron element.

stiffness: Symmetric type, the stiffness matrix of the tetrahedron element.

stress: Symmetric type, the Cauchy stress tensor of the tetrahedron element.

strain: Symmetric type, the strain tensor of the tetrahedron element.

ONSAS.Elements.internal_forces — Method

Return the internal force of a Truss element t formed by an AbstractMaterial m and a an element displacement vector u_e.

ONSAS.Elements.interpolation_matrix — Method

Return the interpolation matrix 𝑀 for a Tetrahedron element t.

ONSAS.Elements.local_dof_symbol — Method

Return local dofs symbols of the AbstractElement e (for linear displacements :u is used) in a vector. Since global degrees of freedom are for the assemble process this function is used to compute the global dofs of the element by extracting the node dofs with the symbol defined by the AbstractElement e.

ONSAS.Elements.local_dof_symbol — Method

Return the local dof symbol of a Tetrahedron element.

ONSAS.Elements.local_dof_symbol — Method

Return the local dof symbol of a Truss element.

ONSAS.Elements.local_dofs — Method

Return local dofs given a vector of local dof symobls. This method extracts all node dofs with the same symbol as localdofsymbol

ONSAS.Elements.nodes — Method

Return the Nodes an AbstractEntity e.

ONSAS.Elements.normal_direction — Method

Return the AbstractFace f normal.

ONSAS.Elements.normal_direction — Method

Return the normal direction n of a TriangularFace element tf.

ONSAS.Elements.num_nodes — Method

Return the number of nodes of element e.

ONSAS.Elements.strain — Method

Return the AbstractElement e strain.

ONSAS.Elements.strain — Method

Return the strain of given Truss element t with a element displacement vector u_e.

ONSAS.Elements.stress — Method

Return the AbstractElement e stress.

ONSAS.Elements.stress — Method

Return the stress of given Truss element t with a element displacement vector u_e.

ONSAS.Elements.volume — Method

Return the Tetrahedron t volume in the reference configuration.

ONSAS.Elements.weights — Method

Return the weights to interpolate a scalar field at the Nodes Dof corresponding to the AbstractElement e.

ONSAS.Elements.weights — Method

Return the interpolation weights of a point p in a Tetrahedron element t.

ONSAS.Utils.apply! — Method

Sets a Vectors of dofs vd to the AbstractNode n assigned to the field s.

ONSAS.Utils.dofs — Method

Return the dofs of an AbstractEntity e.

ONSAS.Utils.dofs — Method

Return AbstractNode n degrees of freedom with symbol s.

ONSAS.Utils.dofs — Method

Return AbstractNode n degrees of freedom.

ONSAS.Utils.dofs — Method

Return the dofs of a Vector ve with AbstractEntityes.

ONSAS.Utils.dofs — Method

Return degrees of freedom for each AbstractNode in a vector of nodes vn.

ONSAS.Utils.label — Method

Return the label of an AbstractEntity e.