Entities

ONSAS.Entities — Module

Module defining an interface, AbstractEntity.

Each entity consists of an object with material and nodes. The following entities are defined:

Entities.

Faces.

ONSAS.Entities.AbstractElement — Type

Abstract supertype for all elements. An AbstractElement object facilitates the process of evaluating: - The internal forces vector and its tangent matrices. - The inertial forces vector and its tangent matrices. - The mechanical stresses and strains.

Abstract Methods

coordinates

dimension

dofs

local_dofs

These methods is a hard contract and must be implemented to define a new element.

local_dof_symbol

label

nodes

weights

Base.:∈

This method is a hard contract and for static analysis must be implemented to define a new element.

internal_forces

This method is a hard contract and for dynamic analysis must be implemented to define a new element.

inertial_forces

Abstract fields

nodes

label

ONSAS.Entities.AbstractEntity — Type

Abstract supertype for all Faces and Elements.

An AbstractEntity object is an entity defined by dofs and node/s with certain coordinates and dimension.

Abstract Methods

create_entity

coordinates

dimension

dofs

Abstract fields

nodes

label

ONSAS.Entities.AbstractFace — Type

Abstract supertype for all elements.

An AbstractFace object facilitates the process of adding boundary conditions on a surface.

Abstract Methods

area

create_entity

coordinates

dofs

label

nodes

normal_direction

Abstract fields

nodes

label

Base.:∈ — Method

Return true if a point p is inside the AbstractElement e.

ONSAS.CrossSections.area — Method

Return the AbstractFace f area.

ONSAS.Entities.create_entity — Method

Return an AbstractEntity given an empty AbstractEntity e and a Vector of Nodes vn.

ONSAS.Entities.cross_section — Method

Return the AbstractElement e cross_section.

ONSAS.Entities.inertial_forces — Method

Return the inertial forces vector of an AbstractElement e. with an AbstractMaterial m

ONSAS.Entities.internal_forces — Method

Return the internal forces vector of an AbstractElement e with an AbstractMaterial m.

ONSAS.Entities.local_dof_symbol — Method

Return local dofs symbols of the AbstractElement e (for linear displacements :u is used) in a vector. Since global degrees of freedom are for the assemble process this function is used to compute the global dofs of the element by extracting the node dofs with the symbol defined by the AbstractElement e.

ONSAS.Entities.local_dofs — Method

Return local dofs given a vector of local dof symobls. This method extracts all node dofs with the same symbol as localdofsymbol

ONSAS.Entities.nodes — Method

Return the Nodes an AbstractEntity e.

ONSAS.Entities.normal_direction — Method

Return the AbstractFace f normal.

ONSAS.Entities.num_nodes — Method

Return the number of nodes of element e.

ONSAS.Entities.strain — Method

Return the AbstractElement e strain.

ONSAS.Entities.stress — Method

Return the AbstractElement e stress.

ONSAS.Entities.volume — Method

Return the volume of the element.

ONSAS.Entities.weights — Method

Return the weights to interpolate a scalar field at the Nodes Dof corresponding to the AbstractElement e.

ONSAS.Nodes.coordinates — Method

Return the AbstractEntity e coordinates.

ONSAS.Nodes.coordinates — Method

Return each AbstractFace coordinates in a Vector of Faces vf.

ONSAS.Nodes.dimension — Method

Return the AbstractEntity dimension.

ONSAS.Utils.dofs — Method

Return the dofs of an AbstractEntity e.

ONSAS.Utils.dofs — Method

Return the dofs of a Vector ve with AbstractEntityes.

ONSAS.Utils.label — Method

Return the label of an AbstractEntity e.