Nodes

ONSAS.Nodes — Module

Module defining an interface, AbstractNode.

A Node is a point in space with degrees of freedom.

ONSAS.Nodes.AbstractNode — Type

An AbstractNode object is a point in space with degrees of freedom.

Abstract Methods

coordinates

dimension

dofs

Abstract fields

x: coordinates of the node

dofs: mapping from field labels to degrees of freedom

ONSAS.Nodes.Dof — Type

Scalar degree of freedom of the structure.

ONSAS.Nodes.Node — Type

A Node is a point in space. The coordinates of the node are stored using a static array.

ONSAS.Nodes.Node — Method

Node constructor to promote types of coordinates of Point

ONSAS.Nodes.Node — Method

Node constructor with an AbstractVector data type.

ONSAS.Nodes.Node — Method

Node constructor with a NTuple.

ONSAS.Nodes.Point — Type

Construct a point given its coordinates.

It is a type alias for a statically sized array of dimension dim and element type T. Use either as a vararg function, Point(1, 2, 3), or by splatting Point(x...) if x is an (abstract) vector.

Base.getindex — Method

the coordinate at component i from the node.

Base.show — Method

Show the a node.

Base.size — Method

Return a tuple containing the dimensions of the node.

ONSAS.Nodes.coordinates — Method

Return node coordinates.

ONSAS.Nodes.dimension — Method

Return the node dimension (1D, 2D or 3D).

ONSAS.Nodes.set_dofs! — Method

Sets a vector of dofs vd to the node assigned to the field s.

ONSAS.Utils.dofs — Method

Return node degrees of freedom.

ONSAS.Utils.index — Method

Return the dof index of the Dof d

StaticArraysCore.similar_type — Method

Method to inherit from StaticArrays for a node.