Tetrahedrons

ONSAS.Tetrahedrons — Module

Module defining tetrahedron elements.

ONSAS.Tetrahedrons.MINOR_INDICES — Constant

Indices for computing the minors of the interpolation matrix, implemented as a hash table.

ONSAS.Tetrahedrons.∂X∂ζ_1 — Constant

Shape function derivatives.

ONSAS.Tetrahedrons.Tetrahedron — Type

Constructor for a Tetrahedron element without nodes and a label. This function is used to create meshes via GMSH.

ONSAS.Tetrahedrons.Tetrahedron — Type

A Tetrahedron represents a 3D volume element with four nodes.

See [Belytschko] and [Gurtin] for more details.

ONSAS.Tetrahedrons.TetrahedronCache — Type

Contains the cache to compute the element internal forces and stiffness matrix.

Base.:∈ — Method

Checks if a point p is inside a Tetrahedron element t.

ONSAS.Entities.create_entity — Method

Return a Tetrahedron given an empty Tetrahedron t and a Vector of Nodes vn.

ONSAS.Entities.elements_cache — Method

Return an empty Tetrahedron cache.

ONSAS.Entities.internal_forces — Method

Return the internal force of a Tetrahedron element t doted with an AbstractHyperElasticMaterial m + and a an element displacement vector u_e. This function modifies the cache to avoid memory allocations.

ONSAS.Entities.internal_forces — Method

Return the internal forces of a Tetrahedron element t doted with and AbstractMaterial m and an element displacement vector u_e.

ONSAS.Entities.internal_forces — Method

Return the internal force of a Tetrahedron element t doted with an LinearIsotropicMaterial m.

Arguments

material: IsotropicLinearElastic type, the linear elastic material of the tetrahedron element.

element: Tetrahedron type, the tetrahedron element for which internal forces are to be computed.

displacements: AbstractVector type, the nodal displacements of the element.

Return

A 4-tuple containing:

forces: Symmetric type, the internal forces of the tetrahedron element.

stiffness: Symmetric type, the stiffness matrix of the tetrahedron element.

stress: Symmetric type, the Cauchy stress tensor of the tetrahedron element.

strain: Symmetric type, the strain tensor of the tetrahedron element.

ONSAS.Entities.local_dof_symbol — Method

Return the local dof symbol of a Tetrahedron element.

ONSAS.Entities.volume — Method

Return the Tetrahedron t volume in the reference configuration.

ONSAS.Entities.weights — Method

Return the interpolation weights of a point p in a Tetrahedron element t.

ONSAS.Tetrahedrons._coordinates_matrix — Method

Return the reshaped coordinates elem_coords of the tetrahedron element into a 4x3 matrix.

ONSAS.Tetrahedrons._jacobian_mat — Method

Computes Jacobian matrix

ONSAS.Tetrahedrons._shape_functions_derivatives — Function

Return the shape functions derivatives of a Tetrahedron element.

ONSAS.Tetrahedrons._volume — Method

Computes volume element of a tetrahedron given J = det(F).

ONSAS.Tetrahedrons.geometric_stiffness! — Method

Return the geometric stiffness.

ONSAS.Tetrahedrons.interpolation_matrix — Method

Return the interpolation matrix 𝑀 for a Tetrahedron element t.