Trusses

ONSAS.Trusses — Module

Module defining truss elements.

ONSAS.Trusses.AbstractStrainModel — Type

A AbstractStrainModel represents the strain model used.

References

See [ANLE].

ONSAS.Trusses.GreenStrain — Type

A GreenStrain a singleton to use the Green engineering strain model.

The strain is: ϵᵢ = (lₙ² - l₀²) / (2l₀²)

ONSAS.Trusses.RotatedEngineeringStrain — Type

A RotatedEngineeringStrain a singleton to use the engineering strain model.

The strain is: ϵᵢ = (lₙ - l₀) / l₀

ONSAS.Trusses.Truss — Type

Constructor for a Truss element without nodes. This function is used to create meshes via GMSH.

ONSAS.Trusses.Truss — Type

A Truss represents an element composed by two Nodes that transmits axial force only.

References

See [ANLE].

ONSAS.Trusses.Truss — Method

Constructor for a Truss element without nodes a label and strain. This function is used to create meshes via GMSH.

ONSAS.Trusses.Truss — Method

Constructor for a Truss element considering the nodes n₁ and n₂ and the cross-section g and strain model.

ONSAS.Trusses.Truss — Method

Constructor for a Truss element considering the nodes n₁ and n₂ and the cross-section g.

ONSAS.Entities.create_entity — Method

Return a Tetrahedron given an empty Tetrahedron t and a Vector of Nodes vn.

ONSAS.Entities.cross_section — Method

Return the cross-section of a Truss element t.

ONSAS.Entities.internal_forces — Method

Return the internal force of a Truss element t formed by an AbstractHyperElasticMaterial m an element displacement vector u_e and GreenStrain.

ONSAS.Entities.internal_forces — Method

Return the internal force of a Truss element t formed by an AbstractHyperElasticMaterial m an element displacement vector u_e and RotatedEngineeringStrain.

ONSAS.Entities.local_dof_symbol — Method

Return the local dof symbol of a Truss element.

ONSAS.Entities.strain — Method

Return the strain of given Truss element t with a element displacement vector u_e.

ONSAS.Entities.stress — Method

Return the stress of given Truss element t with a element displacement vector u_e.

ONSAS.Trusses._X_rows — Method

Return auxiliar matrices with the element coordinates at the deformed and reference configurations.

ONSAS.Trusses._aux_b — Method

Return auxiliar vectors bref and bdef of a truss element.

ONSAS.Trusses._aux_matrices — Method

Return

ONSAS.Trusses._lengths — Method

Return deformed and reference lengths of a truss element.

ONSAS.Trusses._strain — Method

Return the GreenStrain for a given reference and deformed length l_ini and the deformed length l_def.

ONSAS.Trusses._strain — Method

Return the RotatedEngineeringStrain for a given reference and deformed length l_ini and the deformed length l_def.

ONSAS.Trusses.strain_model — Method

Return the strain model used