Gmsh

ONSAS.Gmsh — Module

Module defining meshes created in GMSH software.

ONSAS.Gmsh.MshFile — Type

Collection of Nodes.

ONSAS.Gmsh.MshFile — Method

Constructor for ab MSH file data type with a file name.

ONSAS.Entities.nodes — Method

Return the Nodess from the MshFile.

ONSAS.Gmsh._getlabels — Method

Return material, entities and boundary conditions labels defined in physical names.

ONSAS.Gmsh.bc_label — Method

Return boundary conditions label applied to the indexed entity

ONSAS.Gmsh.bc_label — Method

Return boundary conditions labels defined in the MSH file.

ONSAS.Gmsh.entity_label — Method

Return entities label for the for the indexed entity.

ONSAS.Gmsh.entity_label — Method

Return entities labels defined in the mesh file.

ONSAS.Gmsh.gmsh_println — Method

Return the as variables prints from the REPL.

ONSAS.Gmsh.material_label — Method

Return material label for an indexed index.

ONSAS.Gmsh.material_label — Method

Return material labels defined in the MSH file.

ONSAS.Gmsh.physical_index — Method

Return physical_index defined in the mesh file for an indexed entity .

ONSAS.Gmsh.physical_index — Method

Return physical indexes defined in the MSH file.

ONSAS.Meshes.connectivity — Method

Return the connectivity defined in the MSH file.

ONSAS.Nodes.dimension — Method

Return the mesh dimension defined in the MSH file.