AbstractLinearElasticMaterial Interface

This section describes the interfaces for different linear elastic material types.

ONSAS.Materials.LinearElasticMaterials.AbstractLinearElasticMaterial — Type

Abstract supertype for all elastic material models.

An AbstractLinearElasticMaterial object facilitates the process of using elastic materials.

Common methods:

elastic_modulus

poisson_ratio

shear_modulus

lamé_parameters

Common fields:

label

ρ(density)

ONSAS.Materials.LinearElasticMaterials.IsotropicLinearElastic — Type

IsotropicLinearElastic material struct.

Fields:

E – Elasticity modulus.

ν – Poisson's ratio.

ρ – density (nothing for static cases).

label – material label.

See this ref.

ONSAS.Materials.LinearElasticMaterials.IsotropicLinearElastic — Method

Material IsotropicLinearElastic constructor with no density parameter ρ.

ONSAS.Materials.LinearElasticMaterials.IsotropicLinearElastic — Method

Material IsotropicLinearElastic constructor with Lamé parameters λ, G and density ρ. See this ref.

ONSAS.Materials.LinearElasticMaterials.bulk_modulus — Method

Return the bulk modulus K for an AbstractLinearElasticMaterial material m.

ONSAS.Materials.LinearElasticMaterials.bulk_modulus — Method

Return the bulk modulus K from a IsotropicLinearElastic material m.

ONSAS.Materials.LinearElasticMaterials.cauchy_stress — Method

Return the cauchy stress tensor σ and the constitutive driver ∂σ∂ϵ considering a IsotropicLinearElastic material m.

ONSAS.Materials.LinearElasticMaterials.cauchy_stress — Method

Return the cauchy stress tensor σ and the constitutive driver ∂σ∂ϵ considering a IsotropicLinearElastic material m.

ONSAS.Materials.LinearElasticMaterials.elasticity_modulus — Method

Return the Elasticity modulus E form an AbstractLinearElasticMaterial material m.

ONSAS.Materials.LinearElasticMaterials.elasticity_modulus — Method

Return the Elasticity modulus E from a IsotropicLinearElastic material m.

ONSAS.Materials.LinearElasticMaterials.lame_parameters — Method

Return Lamé parameters λ and G from an AbstractLinearElasticMaterial material m.

ONSAS.Materials.LinearElasticMaterials.lame_parameters — Method

Return Lamé parameters λ and G from a IsotropicLinearElastic material m.

ONSAS.Materials.LinearElasticMaterials.poisson_ratio — Method

Return the Poisson's ratio ν form an AbstractLinearElasticMaterial material m.

ONSAS.Materials.LinearElasticMaterials.poisson_ratio — Method

Return the Poisson's ratio ν from a IsotropicLinearElastic material m.

ONSAS.Materials.LinearElasticMaterials.shear_modulus — Method

Return shear modulus Gfrom anAbstractLinearElasticMaterialmaterialm`.

ONSAS.Materials.LinearElasticMaterials.shear_modulus — Method

Return the shear modulus G from a IsotropicLinearElastic material m.