LinearStaticAnalysis Interface

This section describes the interfaces to handle with linear static analysis.

ONSAS.StructuralAnalyses.StaticAnalyses.LinearStaticAnalyses.LinearStaticAnalysis — Type

Constructor for LinearStaticAnalysis given a final time (or load factor) t₁ and the number of steps NSTEPS.

ONSAS.StructuralAnalyses.StaticAnalyses.LinearStaticAnalyses.LinearStaticAnalysis — Type

LinearStaticAnalysis struct. A LinearStaticAnalysis is a collection of parameters for defining the static analysis of the structure. In the static analysis, the structure is analyzed at a given load factor (this variable is analog to time). As this analysis is linear the stiffness of the structure remains constant at each displacements iteration step.

Fields:

s – stores the structure to be analyzed.

state – stores the structural state.

λᵥ – stores the load factors vector of the analysis

current_step – stores the current load factor step

ONSAS.StructuralSolvers._solve — Method

Solves an LinearStaticAnalysis sa.

ONSAS.StructuralSolvers._step! — Method

Computes ΔU for solving the LinearStaticAnalysis.