Materials

ONSAS.Materials — Module

Module defining the materials implemented. Overall, each materials consists of a data type with a label, parameters and a density into its fields.

ONSAS.Materials.AbstractMaterial — Type

Abstract supertype for all material models.

An AbstractMaterial object facilitates the process of defining new material models. Different material models leads to different constitutive laws, internal forces and stiffness matrices.

Abstract Methods

parameters

density

label

Abstract fields

label

ρ(density)

ONSAS.Materials.density — Method

Return the AbstractMaterial m density ρ.

ONSAS.Materials.parameters — Method

Return the parameters of type Number in the AbstractMaterial m.

ONSAS.Utils.label — Method

Return the AbstractMaterial m label.