Meshes Interface

This section describes the interfaces for different Meshes types.

ONSAS.Meshes — Module

Module defining a mesh interface, AbstractMesh, and the default implementation Mesh.

Each mesh consists of different kinds of entities, stores as dense arrays:

Nodes.

Elements.

Faces.

In addition to the entities, metadata mapping names to entity indices (EntitySet) can be provided.

ONSAS.Meshes.AbstractMesh — Type

Abstract mesh of dimension dim.

Methods

The following methods are provided by the interface:

dimension

dofs

num_dofs

elements

faces

num_elements

nodes

num_nodes

ONSAS.Meshes.EntitySet — Type

Used to designate node, element and face sets mapping an entity string to a set of indices.

ONSAS.Meshes.Mesh — Type

A Mesh is a collection of Elements, Faces and Nodes that cover the discretized domain, together with Sets of elements and nodes.

Methods

node_set

add_node_to_set!

element_set

add_element_to_set!

face_set

add_face_to_set!

Base.push! — Method

push!(m::AbstractMesh, n::AbstractNode)
push!(m::AbstractMesh, e::AbstractElement)
push!(m::AbstractMesh, f::AbstractFace)

Push a node, element or face to the mesh. Important: these methods do not update the nodes array in the mesh. Used mostly for programmatic construction of meshes via GMSH.

ONSAS.Elements.dimension — Method

Return the dimension of an AbstractMesh.

ONSAS.Elements.nodes — Method

Return the array of nodes.

ONSAS.Elements.nodes — Method

Return a Vector of Nodes with node_set_name in the Mesh m.

ONSAS.Elements.num_nodes — Method

Return the number of nodes.

ONSAS.Meshes._isempty_dofs — Method

Return true if the AbstractMesh m has Dofs defined.

ONSAS.Meshes.add_element_to_set! — Method

Register an element's name in the mesh.

ONSAS.Meshes.add_entity_to_set! — Method

Add an entity to a set, dispatching on the entity type.

ONSAS.Meshes.add_face_to_set! — Method

Register a face's name in the mesh.

ONSAS.Meshes.add_node_to_set! — Method

Register a node's name in the mesh.

ONSAS.Meshes.element — Method

Return the element at index i.

ONSAS.Meshes.element_set — Method

Return the set of elements associated to a given name.

ONSAS.Meshes.element_set — Method

Return the set of named elements.

ONSAS.Meshes.elements — Method

Return the array of elements.

ONSAS.Meshes.elements — Method

Return a Vector of Elements with element_set_name in the Mesh m.

ONSAS.Meshes.face_set — Method

Return the set of faces associated to a given name.

ONSAS.Meshes.face_set — Method

Return the set of named faces.

ONSAS.Meshes.faces — Method

Return the array of faces.

ONSAS.Meshes.faces — Method

Return a Vector of Faces with name in the Mesh m.

ONSAS.Meshes.node_set — Method

Return the set of nodes associated to a given name.

ONSAS.Meshes.node_set — Method

Return the set of named nodes.

ONSAS.Meshes.num_dofs — Method

Return the number of Dofs defined in the AbstractMesh m. This function assumes that Dofs indexes start from Dof(1)

ONSAS.Meshes.num_elements — Method

Return the number of elements.

ONSAS.Meshes.num_faces — Method

Return the number of faces.

ONSAS.Utils.apply! — Method

Adds n dofs_per_node Dofs with dof_symbol to the AbstractMesh m.

ONSAS.Utils.dofs — Method

Return the array of degrees of freedom. The i-th entry contains the Dofs of node with index i in the mesh.

Modules = [ONSAS.Meshes.Gmsh]