Interpolators

ONSAS.Interpolators — Module

Module defining the interface to handle with different interpolators.

ONSAS.Interpolators.AbstractInterpolator — Type

An AbstractInterpolator is an object that stores the weights nodes and elements needed to interpolate a magnitude at a given points.

Common methods:

interpolate

ONSAS.Interpolators.FEMInterpolator — Type

A FEMInterpolator struct stores the weights nodes and elements needed to interpolate the solution at a given set of points. The weight are given by the element's shape functions. The index of each Vector is the index in the Vector of Points where the magnitude is evaluated.

ONSAS.Interpolators.interpolate — Method

Return the interpolated solution magnitude get by a function. The accessor function should be evaluated at each node in the points defined. The accessor function will be evaluated in the nodes defined in the nodetoweights dictionary.

ONSAS.Interpolators.interpolate — Method

Return the interpolated nodal vectorial magnitude in the points defined into the interpolator.

ONSAS.Interpolators.interpolate — Method

Return the interpolated nodal scalar magnitude in the points defined into the interpolator.

ONSAS.Interpolators.node_to_weights — Method

Return a vector of dictionaries that maps the nodes and its corresponding weights to interpolate in points.

ONSAS.Interpolators.points — Method

Return a vector of points where the solution is interpolated. This is the order in which the solution is returned.

ONSAS.Interpolators.points_to_element — Method

Return the element where each point is located.