Searches

ONSAS.Searches — Module

Module defining the methods and types to perform searches in meshes.

ONSAS.Searches.AbstractSearchAlgorithm — Type

An AbstractSearchAlgorithm is a functor to use different searching algorithms.

** Common methods:** evaluate_points_in_elements

ONSAS.Searches.Partition — Type

Serial search algorithm.

ONSAS.Searches.PartitionThreaded — Type

Partitioned and threaded search algorithm.

ONSAS.Searches.Serial — Type

Serial search algorithm.

ONSAS.Searches.Threaded — Type

Multithread search algorithm.

Base.:∈ — Method

Return true if a point is inside a mesh.

ONSAS.Searches.bounding_box — Method

Return a box approximation of a node's vector transforming them to LazySets singletons.

ONSAS.Searches.evaluate_points_in_elements — Method

Return indexes checking if a vector of points is inside a vector of elements with a serial searching algorithm. Two integer vectors are returned, the first one contains the indexes of the points inside the mesh, the second one contains the indexes of the elements containing the points.

ONSAS.Searches.evaluate_points_in_elements — Method

Return indexes checking if a vector of points is inside a vector of elements with a serial searching algorithm. Two vector of integers are returned, the first one contains the indexes of the points inside the mesh, the second one contains the indexes of the elements containing the points.

ONSAS.Searches.evaluate_points_in_mesh — Method

Return indexes checking if points are inside the mesh.

ONSAS.Searches.evaluate_points_in_mesh — Method

Return indexes checking if a vector of points is inside a mesh with a threaded partitioned searching algorithm. Two integers vectors are returned, the first one contains the indexes of the points inside the mesh, the second one contains the indexes of the elements containing the points.

ONSAS.Searches.evaluate_points_in_mesh — Method

Return indexes checking if a vector of points is inside a mesh with a partitioned searching algorithm. Two integers vectors are returned, the first one contains the indexes of the points inside the mesh, the second one contains the indexes of the elements containing the points.