NonLinearStaticAnalyses Interface

This section describes the interfaces to handle with non linear static analysis.

ONSAS.StructuralAnalyses.StaticAnalyses.NonLinearStaticAnalyses.NonLinearStaticAnalysis — Type

Constructor for NonLinearStaticAnalysis given a final time (or load factor) t₁ and the number of steps NSTEPS.

ONSAS.StructuralAnalyses.StaticAnalyses.NonLinearStaticAnalyses.NonLinearStaticAnalysis — Type

NonLinearStaticAnalysis struct. A NonLinearStaticAnalysis is a collection of parameters for defining the static analysis of the structure. In the static analysis, the structure is analyzed at a given load factor (this variable is analog to time). As this analysis is nonlinear the stiffness of the structure is updated at each iteration.

Fields:

s – stores the structure to be analyzed.

state – stores the structural state.

λᵥ – stores the load factors vector of the analysis

current_step – stores the current load factor step

ONSAS.StructuralSolvers._solve — Method

Solves an NonLinearStaticAnalysis sa with an AbstractSolver alg.

ONSAS.StructuralSolvers._step! — Method

Computes ΔU for solving the NonLinearStaticAnalysis sa with a NewtonRaphson method.