AbstractStaticAnalyses Interface

This section describes the interfaces to handle with static analysis.

ONSAS.StructuralAnalyses.StaticAnalyses.AbstractStaticAnalysis — Type

Abstract supertype for all structural analysis.

An AbstractStaticAnalysis object facilitates the process of defining an static analysis to be solved. Time variable for static analysis is used to obtain a load factor value. Of course this abstract type inherits from AbstractStructuralAnalysis type, and extends the following methods:

Common methods:

initial_time

current_time

final_time

load_factors

current_load_factor

_next!

is_done

reset!

_solve

_step!

Common fields

λᵥ – stores the load factors vector.

current_step – stores the current step of the analysis.

state – stores the current state of the analysis.

ONSAS.StructuralAnalyses.StaticAnalyses.StaticState — Type

An StaticState object facilitates the process of storing the relevant static variables of the structure during the displacements iteration.

Fields:

s – stores the structure

ΔUᵏ – stores displacements vector increment.

Uᵏ – stores displacements vector.

Fₑₓₜᵏ – stores external forces vector.

Fᵢₙₜᵏ – stores internal forces vector.

Kₛᵏ – stores the system tangent matrix.

ϵᵏ – stores a vector with strains for each element.

σᵏ – stores a vector with stresses for each element.

assembler – assembler handler object

iter_state – current Δu iteration state

ONSAS.StructuralAnalyses.StaticAnalyses.StaticState — Method

Constructor for StaticState given an AbstractStructure s.

Base.push! — Method

Pushes the current state c_state into the StatesSolution st_sol.

ONSAS.StructuralAnalyses.StaticAnalyses.current_load_factor — Method

Returns the current load factor of an AbstractStaticAnalysis sa.

ONSAS.StructuralAnalyses.StaticAnalyses.load_factors — Method

Returns the final load factor vector of an AbstractStaticAnalysis sa.

ONSAS.StructuralAnalyses._next! — Method

Jumps to the next current load factor defined in the AbstractStaticAnalysis sa.

ONSAS.StructuralAnalyses.current_time — Method

Returns the current load factor of an AbstractStaticAnalysis sa.

ONSAS.StructuralAnalyses.final_time — Method

Returns the final load factor of an AbstractStaticAnalysis sa.

ONSAS.StructuralAnalyses.initial_time — Method

Returns the initial load factor of an AbstractStaticAnalysis sa.

ONSAS.StructuralAnalyses.is_done — Method

Returns true if the AbstractStaticAnalysis sa is completed.

ONSAS.StructuralAnalyses.reset! — Method

Sets the current load factor of the AbstractStaticAnalysis sa to the initial load factor.

ONSAS.StructuralAnalyses.residual_forces — Method

Returns the current residual forces of the StaticState sc.

ONSAS.StructuralAnalyses.tangent_matrix — Method

Returns the current system tangent matrix of the StaticState sc.

ONSAS.StructuralSolvers._assemble! — Method

Assembles the Structure s (internal forces) during the StaticAnalysis sa.

ONSAS.StructuralSolvers._reset! — Method

Resets the StaticState assembled magnitudes.

ONSAS.StructuralSolvers._update! — Method

Updates displacements in the StaticState sc with a displacements increment vector ΔU.