AbstractStructuralAnalyses Interface

This section describes the interfaces to handle with different structural analysis.

ONSAS.StructuralAnalyses — Module

Module defining structural analyses that can be solved. Each structural analysis consists of a data type with an structure to be analyzed, analysis parameters and an specific state which describes the current state of structure in the analysis.

ONSAS.StructuralAnalyses.AbstractStructuralAnalysis — Type

Abstract supertype for all structural analysis.

An AbstractStructuralAnalysis object facilitates the process of defining an structural analysis to be solved.

Common methods:

structure

free_dofs

structural_state

initial_time

current_time

final_time

current_state

current_iteration

_next!

is_done

reset!

ONSAS.StructuralAnalyses.AbstractStructuralState — Type

Abstract supertype to define a new structural state. Common methods:

Accessors:

displacements

Δ_displacements

external_forces

internal_forces

residual_forces!

tangent_matrix

strain

stress

structure

free_dofs

Iteration:

_assemble!

assembler

iteration_residuals

residual_forces_norms

residual_displacements_norms

ONSAS.Assemblers._assemble! — Method

Assembles the element e stiffness matrix matrix K_e into the AbstractState st

ONSAS.Assemblers._assemble! — Method

Assembles the element e internal forces fᵢₙₜ_e into the AbstractState st

ONSAS.Assemblers._assemble! — Method

Assembles the element e stress σₑ and strain ϵₑ into the AbstractState st

ONSAS.Assemblers._end_assemble! — Method

Fill the system tangent matrix in the AbstractStructuralState st once the Assembler object is built.

ONSAS.Entities.internal_forces — Method

Return the current internal forces vector in the AbstractStructuralState st.

ONSAS.Entities.strain — Method

Return strains for each Element in the AbstractStructuralState st.

ONSAS.Entities.stress — Method

Return stresses for each Element in the AbstractStructuralState st.

ONSAS.Solutions.displacements — Method

Return current displacements vector at the current AbstractStructuralState st.

ONSAS.Solutions.external_forces — Method

Return external forces vector in the AbstractStructuralState st.

ONSAS.Solutions.iteration_residuals — Method

Return current ResidualsIterationStep object form an AbstractStructuralState st.

ONSAS.StructuralAnalyses._next! — Method

Increments the time step given the AbstractStructuralAnalysis a and the AbstractStructuralSolver alg.

ONSAS.StructuralAnalyses.assembler — Method

Return the Assembler used in the AbstractStructuralState st.

ONSAS.StructuralAnalyses.current_iteration — Method

Return the current displacements iteration state of the AbstractStructuralAnalysis a.

ONSAS.StructuralAnalyses.current_state — Method

Return the current state of the AbstractStructuralAnalysis a.

ONSAS.StructuralAnalyses.current_time — Method

Return the current time of AbstractStructuralAnalysis a.

ONSAS.StructuralAnalyses.final_time — Method

Return the final time of AbstractStructuralAnalysis a.

ONSAS.StructuralAnalyses.initial_time — Method

Return the initial time of AbstractStructuralAnalysis a.

ONSAS.StructuralAnalyses.is_done — Method

Return true if the AbstractStructuralAnalysis a is completed.

ONSAS.StructuralAnalyses.residual_displacements_norms — Method

Return relative residual displacements for the current AbstractStructuralState st.

ONSAS.StructuralAnalyses.residual_forces! — Method

Return residual forces vector in the AbstractStructuralState st.

ONSAS.StructuralAnalyses.residual_forces_norms — Method

Return relative residual forces for the current AbstractStructuralState st.

ONSAS.StructuralAnalyses.structure — Method

Return analyzed structure in the AbstractStructuralAnalysis a.

ONSAS.StructuralAnalyses.tangent_matrix — Method

Return system tangent matrix in the AbstractStructuralState st.

ONSAS.StructuralAnalyses.Δ_displacements — Method

Return current displacements increment vector at the current AbstractStructuralState st.

ONSAS.StructuralModel.free_dofs — Method

Return free Dofs of the structure in the AbstractStructuralState st.

ONSAS.StructuralSolvers._update! — Method

Updates the AbstractStructuralState st during the displacements iteration.

ONSAS.StructuralSolvers.reset! — Method

Resets the AbstractStructuralState assembled magnitudes before starting a new assembly.

ONSAS.StructuralSolvers.reset! — Method

Rests the AbstractStructuralAnalysis sa (sets the current time to the initial time).

ONSAS.Utils.apply! — Method

Apply an AbstractNeumannBoundaryCondition lbc into the structural analysis sa at the current analysis time t

ONSAS.Utils.apply! — Method

Apply a vector of load boundary conditions to the structure s