StaticStates

ONSAS.StaticStates — Module

Module defining static states. Each static state defines the structural state of an structure into a StaticAnalysis. This state contains external and internal forces, displacements, stresses and strains.

ONSAS.StaticStates.StaticState — Type

Default constructor for static state given an structure and iteration state.

ONSAS.StaticStates.StaticState — Type

Stores the relevant static variables of the structure during the displacements iteration.

ONSAS.Assemblers.reset! — Method

Reset the static state assembled magnitudes and the iteration state.

ONSAS.StructuralAnalyses.residual_forces! — Method

Update and return the current residual forces of the static state.

ONSAS.StructuralAnalyses.tangent_matrix — Method

Return the current system tangent matrix form the static state .