Structural Analysis Interface

This section describes the interfaces to handle with different structural analysis.

ONSAS.StructuralAnalyses — Module

Module defining structural analyses that can be solved. Each structural analysis consists of a data type with an structure to be analyzed, analysis parameters and an specific state which describes the current state of structure in the analysis.

ONSAS.StructuralAnalyses.AbstractStructuralAnalysis — Type

Abstract supertype for all structural analysis.

An AbstractStructuralAnalysis object facilitates the process of defining an structural analysis to be solved.

Common methods:

structure

free_dofs

structural_state

initial_time

current_time

final_time

current_state

current_iteration

_next!

is_done

ONSAS.StructuralAnalyses.AbstractStructuralState — Type

Abstract supertype to define a new structural state. Common methods:

Accessors:

displacements

Δ_displacements

external_forces

internal_forces

residual_forces

tangent_matrix

strain

stress

structure

free_dofs

Iteration:

_assemble!

assembler

iteration_residuals

residual_forces_norms

residual_displacements_norms

ONSAS.Elements.internal_forces — Method

Returns the current internal forces vector in the AbstractStructuralState st.

ONSAS.Elements.strain — Method

Returns strains for each Element in the AbstractStructuralState st.

ONSAS.Elements.stress — Method

Returns stresses for each Element in the AbstractStructuralState st.

ONSAS.StructuralAnalyses._apply! — Method

Applies an AbstractLoadBoundaryCondition lbc into the structural analysis sa at the current analysis time t

ONSAS.StructuralAnalyses._apply! — Method

Applies a vector of load boundary conditions to the structure s

ONSAS.StructuralAnalyses._next! — Method

Increments the time step given the AbstractStructuralAnalysis a and the AbstractStructuralSolver alg.

ONSAS.StructuralAnalyses.assembler — Method

Returns the Assembler used in the AbstractStructuralState st.

ONSAS.StructuralAnalyses.current_iteration — Method

Returns the current displacements iteration state of the AbstractStructuralAnalysis a.

ONSAS.StructuralAnalyses.current_state — Method

Returns the current state of the AbstractStructuralAnalysis a.

ONSAS.StructuralAnalyses.current_time — Method

Returns the current time of AbstractStructuralAnalysis a.

ONSAS.StructuralAnalyses.final_time — Method

Returns the final time of AbstractStructuralAnalysis a.

ONSAS.StructuralAnalyses.initial_time — Method

Returns the initial time of AbstractStructuralAnalysis a.

ONSAS.StructuralAnalyses.is_done — Method

Returns true if the AbstractStructuralAnalysis a is completed.

ONSAS.StructuralAnalyses.residual_displacements_norms — Method

Returns relative residual displacements for the current AbstractStructuralState st.

ONSAS.StructuralAnalyses.residual_forces — Method

Returns residual forces vector in the AbstractStructuralState st.

ONSAS.StructuralAnalyses.residual_forces_norms — Method

Returns relative residual forces for the current AbstractStructuralState st.

ONSAS.StructuralAnalyses.structure — Method

Returns analyzed structure in the AbstractStructuralAnalysis a.

ONSAS.StructuralAnalyses.structure — Method

Returns the structure of the AbstractStructuralState st.

ONSAS.StructuralAnalyses.tangent_matrix — Method

Returns system tangent matrix in the AbstractStructuralState st.

ONSAS.StructuralAnalyses.Δ_displacements — Method

Returns current displacements increment vector at the current AbstractStructuralState st.

ONSAS.StructuralModel.free_dofs — Method

Returns free Dofs of the structure in the AbstractStructuralState st.

ONSAS.StructuralSolvers._assemble! — Method

Assembles the element e stiffness matrix matrix K_e into the AbstractState st

ONSAS.StructuralSolvers._assemble! — Method

Assembles the element e internal forces fᵢₙₜ_e into the AbstractState st

ONSAS.StructuralSolvers._assemble! — Method

Assembles the element e stress σₑ and strain ϵₑ into the AbstractState st

ONSAS.StructuralSolvers._end_assemble! — Method

Fill the system tangent matrix in the AbstractStructuralState st once the Assembler object is built.

ONSAS.StructuralSolvers._reset! — Method

Resets the AbstractStructuralState assembled magnitudes before starting a new assembly.

ONSAS.StructuralSolvers._update! — Method

Updates the AbstractStructuralState st during the displacements iteration.

ONSAS.StructuralSolvers.displacements — Method

Returns current displacements vector at the current AbstractStructuralState st.

ONSAS.StructuralSolvers.external_forces — Method

Returns external forces vector in the AbstractStructuralState st.

ONSAS.StructuralSolvers.iteration_residuals — Method

Returns current ResidualsIterationStep object form an AbstractStructuralState st.