Structural Model Interface

This section describes the interfaces for defining structural models.

ONSAS.StructuralModel — Module

Module defining structure entities interface. Each Structure consists of a data type with materials, boundary conditions, mesh and a vector of free dofs.

ONSAS.StructuralModel.AbstractStructure — Type

Abstract supertype to define a new structure. An AbstractStructure object facilitates the process of assigning materials, elements, and boundary conditions to the mesh. Moreover this struct is used to solve an specific structural analysis.

Common methods:

Mesh:

dofs

num_dofs

free_dofs

num_free_dofs

nodes

num_nodes

faces

num_faces

elements

num_elements

mesh

Boundary conditions:

boundary_conditions

displacement_bcs

load_bcs

node_bcs

element_bcs

Materials:

materials

ONSAS.Entities.nodes — Method

Return a Vector of Nodes defined in the AbstractStructure s.

ONSAS.Entities.num_nodes — Method

Return the number of Nodes of the AbstractStructure s

ONSAS.Handlers.mesh — Method

Return the Mesh of the AbstractStructure s

ONSAS.Meshes.elements — Method

Return the Elements of the AbstractStructure s

ONSAS.Meshes.faces — Method

Return the Faces of the AbstractStructure s

ONSAS.Meshes.num_dofs — Method

Return the number of Dofs of the AbstractStructure s

ONSAS.Meshes.num_elements — Method

Return the number of Elements of the AbstractStructure s

ONSAS.Meshes.num_faces — Method

Return the number of Faces of the AbstractStructure s

ONSAS.StructuralModel.boundary_conditions — Method

Return the StructuralBoundaryCondition of the AbstractStructure s

ONSAS.StructuralModel.displacement_bcs — Method

Return the DisplacementBoundaryConditions of the AbstractStructure s

ONSAS.StructuralModel.element_bcs — Method

Return the AbstractBoundaryConditions imposed to Elements in the AbstractStructure s

ONSAS.StructuralModel.face_bcs — Method

Return the AbstractBoundaryConditions imposed to Faces in the AbstractStructure s

ONSAS.StructuralModel.free_dofs — Method

Return free Dofs of the structure s

ONSAS.StructuralModel.load_bcs — Method

Return the AbstractLoadBoundaryConditions of the AbstractStructure s

ONSAS.StructuralModel.materials — Method

Return the StructuralMaterials of the AbstractStructure s

ONSAS.StructuralModel.node_bcs — Method

Return the AbstractBoundaryConditions imposed to Nodes in the AbstractStructure s

ONSAS.StructuralModel.num_free_dofs — Method

Return the number of free Dofs of the AbstractStructure s

ONSAS.Utils.dofs — Method

Return the Dofs of the AbstractStructure s