StructuralBoundaryConditions

ONSAS.StructuralBoundaryConditions — Module

Module defining the structural boundary conditions assigned to each entitiy. This data type includes:

AbstractBoundaryConditions assigned to Nodes

AbstractBoundaryConditions assigned to Faces

AbstractBoundaryConditions assigned to Elements

ONSAS.StructuralBoundaryConditions.StructuralBoundaryCondition — Type

Structural boundary conditions. A StructuralBoundaryCondition is a collection of BoundaryConditions defining the boundary conditions of the structure.

ONSAS.StructuralBoundaryConditions.StructuralBoundaryCondition — Method

Constructor for empty StructuralBoundaryCondition with a Vector of AbstractBoundaryConditions vbc.

Base.delete! — Method

Delete a boundary condition.

Base.getindex — Method

Return the boundary conditions applied to an element.

Base.getindex — Method

Return the boundary conditions applied to a face.

Base.getindex — Method

Return the boundary conditions applied to a node.

Base.getindex — Method

Return the bc with the label l in the structural boundary conditions.

Base.getindex — Method

Return a the entities where the boundary condition is applied.

Base.getindex — Method

Return a the boundary conditions with the label l in the structural boundary conditions.

Base.insert! — Method

Insert a boundary condition.

Base.replace! — Function

Replace the boundary condition with the same label as the new.

ONSAS.BoundaryConditions.apply — Method

Return a vector of dofs and load values applied by the boundary condition at time t.

ONSAS.BoundaryConditions.apply — Method

Return the dofs that the boundary condition fixes.

ONSAS.StructuralBoundaryConditions.all_bcs — Method

Return all the boundary conditions applied to the structure.

ONSAS.StructuralBoundaryConditions.apply_element_bcs! — Method

Apply element boundary conditions to the mesh. Assign entities with the same label in the boundary conditions set.

ONSAS.StructuralBoundaryConditions.apply_face_bcs! — Method

Apply face boundary conditions to the mesh. Assign entities with the same label in the boundary conditions set.

ONSAS.StructuralBoundaryConditions.apply_node_bcs! — Method

Apply node boundary conditions to the mesh. Assign entities with the same label in the boundary conditions set.

ONSAS.StructuralBoundaryConditions.displacement_bcs — Method

Return a displacement boundary conditions.

ONSAS.StructuralBoundaryConditions.element_bcs — Method

Return the element boundary conditions.

ONSAS.StructuralBoundaryConditions.face_bcs — Method

Return the dictionary of face boundary conditions.

ONSAS.StructuralBoundaryConditions.fixed_dof_bcs — Method

Return fixed dofs boundary conditions.

ONSAS.StructuralBoundaryConditions.load_bcs — Method

Return the load boundary conditions.

ONSAS.StructuralBoundaryConditions.node_bcs — Method

Return the dictionary of node boundary conditions.

ONSAS.Utils.apply! — Method

Apply all the boundary conditions to the mesh.