StructuralEntities

ONSAS.StructuralEntities — Module

Module defining structural entities properties to then construct an structure. This used to create a mesh with GMSH, and the structural entities contain the ONSAS element type to the posterior conversion to the ONSAS mesh type.

ONSAS.Meshes.Mesh — Method

Return the Mesh given an MshFile msh_file and a StructuralEntity entities.

ONSAS.StructuralEntities.StructuralEntity — Type

Struct used to define meshes via GMSH. Consists of a collection of element and face types, each assigned to a corresponding vector of Faces and Elements.

ONSAS.StructuralEntities.StructuralEntity — Method

Constructor for an empty StructuralEntity with a Vector of Elements velems and Faces vfaces.

Base.getindex — Method

Return the Entity with the label l in the StructuralEntity entities.

ONSAS.StructuralEntities.all_entities — Method

Return all Entitys defined into StructuralEntity.

ONSAS.StructuralEntities.elem_types — Method

Return the Vector of Element types defined in the StructuralEntity entities.

ONSAS.StructuralEntities.elem_types_to_elements — Method

Return a Dictionary with Element types as keys and the corresponding Elements as values.

ONSAS.StructuralEntities.face_types — Method

Return the Vector of Face types defined in the StructuralEntity entities.

ONSAS.StructuralEntities.face_types_to_faces — Method

Return a Dictionary with Face types as keys and the corresponding Faces as values.