StructuralMaterials

ONSAS.StructuralMaterials — Module

"Module defining structural material. This used to assign materials to elements when creating the mesh to the corresponding material type."

ONSAS.StructuralMaterials.MEPair — Type

Material associated to an array of elements.

ONSAS.StructuralMaterials.StructuralMaterial — Type

A StructuralMaterial is a collection of Materials and Elements assigning materials to a vector of elements.

ONSAS.StructuralMaterials.StructuralMaterial — Method

Constructor for empty StructuralMaterial with a Vector of materials vmats. This will assign an empty Vector of Elements to each material.

Base.delete! — Method

Delete the Material m from the StructuralMaterial sm.

Base.getindex — Method

Return the Material mapped with the label l.

Base.getindex — Method

Return the Vector of Material of the element e.

Base.getindex — Method

Return the Vector of Elements that are conformed by the Materialm`.

Base.insert! — Method

Insert the Material m into the StructuralMaterial sm.

Base.pairs — Method

Return Pairs of Material and Element in the StructuralMaterial sm.

Base.replace! — Function

Replace the AbstractMaterial with the label l for the new material new_m in the StructuralMaterial sm. The previous material Elements are assigned to the new.

ONSAS.StructuralMaterials._element_material_is_unique — Method

Checks that each Element has a single Material in the dictionary mat_dict.

ONSAS.StructuralMaterials.element_materials — Method

Return a Dictionary with Materials as keys and the corresponding Elements as values.

ONSAS.Utils.apply! — Method

Apply the StructuralBoundaryCondition to the AbstractMesh m. For this is required sets into the Mesh and the corresponding boundary condition labels declared in bcs.