Structural Solvers Interface

This section describes the interfaces for structural solvers.

ONSAS.StructuralSolvers — Module

Module defining structural solvers that can be used to solved different analyses. Each solver consists of a data type with a convergence criterion and a the iteration status. A _step! method is used to perform a single iteration step.

ONSAS.StructuralSolvers.AbstractConvergenceCriterion — Type

Abstract supertype for all convergence criterion.

ONSAS.StructuralSolvers.AbstractSolution — Type

Abstract supertype for all structural analysis solutions.

ONSAS.StructuralSolvers.AbstractSolver — Type

Abstract supertype for all direct integration methods.

ONSAS.StructuralSolvers.Assembler — Type

Assembler struct This struct stores column indexes, row indexes and values for the assemble process.

Fields:

I – column indexes

J – row indexes

V – values

ONSAS.StructuralSolvers.Assembler — Method

Constructor of an Assembler with size of the sparse matrix N .

ONSAS.StructuralSolvers.ConvergenceSettings — Type

ConvergenceSettings struct. Facilitates the process of defining and checking numerical convergence.

Fields:

rel_disp_tol – Relative displacement tolerance.

rel_force_tol – Relative residual force tolerance.

max_iter – Maximum number of iterations.

ONSAS.StructuralSolvers.MaxIterCriterion — Type

MaxIterCriterion criteria indicates that the maximum number of iterations has been reached.

ONSAS.StructuralSolvers.NewtonRaphson — Type

Newton-Raphson solver struct.

Fields:

tols – Numerical tolerances

ONSAS.StructuralSolvers.NewtonRaphson — Method

Constructor for NewtonRaphson solver with default tolerances.

ONSAS.StructuralSolvers.NotConvergedYet — Type

NotConvergedYet indicates that the current iteration has not converged.

ONSAS.StructuralSolvers.ResidualForceCriterion — Type

ResidualForceCriterion convergence criterion.

ONSAS.StructuralSolvers.ResidualsIterationStep — Type

ResidualsIterationStep struct. Stores the convergence information of at the current iteration step.

Fields:

Δu_norm – Norm of the displacement increment. Δr_norm – Norm of the residual force increment. Δu_rel – Relative norm of the displacement increment. Δr_rel – Relative norm of the residual force increment. n_iter – Current iteration number.

ONSAS.StructuralSolvers.StatesSolution — Type

Structural analysis solution containing the full history of states. Common methods:

displacements

external_forces

internal_forces

stresses

strains

Common fields:

ONSAS.StructuralSolvers.ΔUCriterion — Type

ΔUCriterion indicates displacements increment convergence criterion.

ONSAS.StructuralSolvers.ΔU_and_ResidualForce_Criteria — Type

ΔU_and_ResidualForce_Criteria convergence criterion indicates that both ΔU and residual forces converged .

ONSAS.StructuralSolvers._assemble! — Method

Assembles the element matrix Ke into the Assembler struct a.

ONSAS.StructuralSolvers._assemble! — Method

Assembles the matrix Ke into Assembler a according to the dofs specified by row_dof_indexes and col_dof_indexes.

ONSAS.StructuralSolvers._end_assemble! — Method

Inserts an the Assembler object a into the tangent system matrix K_sys.

ONSAS.StructuralSolvers._init — Method

Returns the initialized analysis.

ONSAS.StructuralSolvers._reset! — Method

Empties the Assembler object a. This is useful to reuse the same assembler.

ONSAS.StructuralSolvers._reset! — Method

Sets the ResidualsIterationStep i_step iteration step to 0.

ONSAS.StructuralSolvers._solve — Method

Internal solve function to be overloaded by each analysis.

ONSAS.StructuralSolvers._step! — Method

Increments a ResidualsIterationStep i_step by 1.

ONSAS.StructuralSolvers._step! — Method

Computes a step in time on the analysis considering the numerical AbstractSolver solver.

ONSAS.StructuralSolvers._update! — Method

Updates the ResidualsIterationStep i_step current convergence criterion.

ONSAS.StructuralSolvers._update! — Method

Updates the ResidualsIterationStep i_step iteration step with the current values of the displacement residuals ΔU_norm and ΔU_rel, the norm of the residual forces increment Δr_norm, and the relative Δr_rel.

ONSAS.StructuralSolvers.analysis — Method

Returns the AbstractAnalysis solved.

ONSAS.StructuralSolvers.criterion — Method

Returns the current criterion of the ResidualsIterationStep i_step.

ONSAS.StructuralSolvers.displacement_tol — Method

Returns displacements tolerance set in the ConvergenceSettings tol.

ONSAS.StructuralSolvers.displacement_tol — Method

Returns the current absolute and relative residual forces norm of the ResidualsIterationStep i_step.

ONSAS.StructuralSolvers.end_assemble — Method

Returns an assembled AbstractSparseMatrix from the Assembler object a.

ONSAS.StructuralSolvers.isconverged! — Method

Updates the ResidualsIterationStep i_step given a ConvergenceSettings cs.

ONSAS.StructuralSolvers.iter — Method

Returns the current iteration number of the ResidualsIterationStep i_step.

ONSAS.StructuralSolvers.iteration_residuals — Method

Returns the IterationResiduals object at every time step.

ONSAS.StructuralSolvers.max_iter_tol — Method

Returns the maximum number of iterations set in the ConvergenceSettings tol.

ONSAS.StructuralSolvers.residual_forces_tol — Method

Returns residual forces tolerance set in the ConvergenceSettings tol.

ONSAS.StructuralSolvers.residual_forces_tol — Method

Returns the current absolute and relative residual forces norm of the ResidualsIterationStep i_step.

ONSAS.StructuralSolvers.solve — Method

Solve an structural analysis problem.

Input

analysis – structural analysis problem

alg – structural algorithm to solve the problem

Output

A solution structure (AbstractSolution) that holds the result and the algorithm used to obtain it.

ONSAS.StructuralSolvers.solver — Method

Returns the AbstractSolver solved.

ONSAS.StructuralSolvers.states — Method

Returns the solved AbstractStrcturalStates.

ONSAS.StructuralSolvers.step_size — Method

Returns the step size set to the AbstractSolver solver .

ONSAS.StructuralSolvers.tolerances — Method

Returns the numerical tolerances set to the AbstractSolver solver.

ONSAS.StructuralSolvers.tolerances — Method

Returns NewtonRaphson solver tolerances.