Solutions

ONSAS.Solutions — Module

Module defining solutions interface, AbstractSolution.

Each solution can at least access to displacements and external forces. More complete solutions can include different forces vector, stresses and strains.

Of course the solution contains on the analysis and solver used to solve the problem.

ONSAS.Solutions.AbstractSolution — Type

Abstract supertype for all structural analysis solutions.

Abstract Methods

displacements

external_forces

internal_forces

stress

strain

Abstract fields

analysis

solver

ONSAS.Solutions.StatesSolution — Type

Solution that stores all intermediate arrays during the analysis.

Base.show — Method

Show the states solution.

ONSAS.Entities.internal_forces — Method

Return the internal force component at the points in the point eval handler.

ONSAS.Entities.internal_forces — Method

Return the internal forces solution at the points in the point eval handler.

ONSAS.Entities.strain — Method

Return the strains at the elements where the points are located in the point eval handler.

ONSAS.Entities.stress — Method

Return the stress at the elements where the points are located in the point eval handler.

ONSAS.Solutions._print_table — Method

Print the solution table

ONSAS.Solutions.analysis — Method

Return the analysis solved.

ONSAS.Solutions.deformed_node_positions — Method

Return deformed nodes of the mesh at time index.

ONSAS.Solutions.solver — Method

Return the solver used to solve the analysis.

ONSAS.Solutions.states — Method

Return the solved states.

ONSAS.StructuralAnalyses.displacements — Method

Return the displacements component in the point evaluator handler.

ONSAS.StructuralAnalyses.displacements — Method

Return the displacements solution at the points in the point evaluator handler.

ONSAS.StructuralAnalyses.external_forces — Method

Return the external force component at the points in the point eval handler.

ONSAS.StructuralAnalyses.external_forces — Method

Return the external forces at the points in the point eval handler.

ONSAS.StructuralAnalyses.iteration_residuals — Method

Return the residuals iteration object at every time step.

ONSAS.StructuralSolvers.criterion — Method

Return the convergence criterion at each time step.

ONSAS.StructuralSolvers.displacement_tol — Method

Return the final residual displacement at each time step.

ONSAS.StructuralSolvers.iterations — Method

Return the number of iterations at each time step.

ONSAS.StructuralSolvers.residual_forces_tol — Method

Return the final residuals forces at each time step.