Structural Model Interface

This section describes the interfaces for defining structural models.

ONSAS.Structures — Module

Module defining structure entities interface. Each Structure consists of a data type with materials, boundary conditions, mesh and a vector of free dofs.

ONSAS.Handlers.PointEvalHandler — Method

Constructor of a point eval handler from a structure and a vector of points .

ONSAS.Structures.AbstractStructure — Type

Abstract supertype to define a new structure. An AbstractStructure object facilitates the process of assigning materials, elements, and boundary conditions to the mesh. Moreover this struct is used to solve an specific structural analysis.

Abstract Methods

Mesh:

dofs

num_dofs

free_dofs

num_free_dofs

nodes

num_nodes

faces

num_faces

elements

num_elements

mesh

Boundary conditions:

boundary_conditions

displacement_bcs

load_bcs

node_bcs

element_bcs

Materials:

materials

ONSAS.Structures.Structure — Type

Constructor for a structure given a mesh file, and their corresponding structural materials, structural elements and structural boundary conditions.

ONSAS.Structures.Structure — Type

An Structure object facilitates the process of assembling and creating the structural analysis.

ONSAS.Structures.Structure — Method

Constructor for a structure with mesh, materials and boundary conditions.

Base.replace! — Function

Replace a material with the a label for the new material in the structure. The previous material element's are assigned to the new.

ONSAS.Entities.nodes — Method

Return the structure nodes.

ONSAS.Entities.num_nodes — Method

Return the structure number of nodes.

ONSAS.Handlers.mesh — Method

Return the structure mesh.

ONSAS.Meshes.elements — Method

Return the structure elements.

ONSAS.Meshes.faces — Method

Return the structure faces.

ONSAS.Meshes.num_dofs — Method

Return structure dofs.

ONSAS.Meshes.num_elements — Method

Return the structure number of elements.

ONSAS.Meshes.num_faces — Method

Return the number of Faces of the AbstractStructure s

ONSAS.Structures.boundary_conditions — Method

Return the structure boundary conditions.

ONSAS.Structures.free_dofs — Method

Return the structure free dofs.

ONSAS.Structures.materials — Method

Return the structure materials.

ONSAS.Structures.num_free_dofs — Method

Return the number of free dofs of the structure.

ONSAS.Utils.dofs — Method

Return the structure dofs.