Utils

ONSAS.Utils.Density — Type

Physical paramater defining density of a material (nothing reserved for static cases).

ONSAS.Utils.Label — Type

Used to assign labels to geometric or physical entities.

ONSAS.Utils.NO_LABEL — Constant

Label to design an entity without assigned label.

ONSAS.Utils.Field — Type

Type alias used for field labels as degree-of-freedom keys such as :T or :θ.

ONSAS.Utils.ScalarWrapper — Type

Scalar mutable struct to avoid making mutable larger structs.

ONSAS.Utils._unwrap — Function

Unwraps the object fields.

ONSAS.Utils.apply! — Function

Apply one object to the other.

ONSAS.Utils.dofs — Function

Return the dofs of an object

ONSAS.Utils.eye — Function

Return an eye matrix of size m and type T.

ONSAS.Utils.index — Function

Return the index of an object

ONSAS.Utils.label — Function

Return the label of an object.

ONSAS.Utils.row_vector — Method

Transforms a vector of vectors into a 1D row vector.

ONSAS.Utils.voigt — Function

Return the tensor 𝕋 in Voigt notation.

ONSAS.Utils.@debugtime — Macro

Execute an expression returning the result and printing the elapsed time inside a @debug statement.