Elements Interface

This section describes the interfaces for different element types.

ONSAS.Elements — Module

Module defining element formulations implemented. Each element consists of a data type with the geometric properties such as nodes, cross-sections and a label into its fields.

ONSAS.Elements.AbstractElement — Type

Abstract supertype for all elements. An AbstractElement object facilitates the process of evaluating: - The internal forces vector and its tangent matrices. - The inertial forces vector and its tangent matrices. - The mechanical stresses and strains.

Common methods:

coordinates

dimension

dofs

local_dofs

These methods is a hard contract and must be implemented to define a new element.

local_dof_symbol

label

nodes

weights

Base.:∈

This method is a hard contract and for static analysis must be implemented to define a new element.

internal_forces

This method is a hard contract and for dynamic analysis must be implemented to define a new element.

inertial_forces

Common fields:

nodes

label

ONSAS.Elements.AbstractEntity — Type

Abstract supertype for all Faces and Elements.

An AbstractEntity object is an entity defined by dofs and node/s with certain coordinates and dimension.

Common methods:

create_entity

coordinates

dimension

dofs

Common fields:

nodes

label

ONSAS.Elements.AbstractFace — Type

Abstract supertype for all elements.

An AbstractFace object facilitates the process of adding boundary conditions on a surface.

Common methods:

area

create_entity

coordinates

dofs

label

nodes

normal_direction

Common fields:

nodes

label

ONSAS.Elements.AbstractNode — Type

Abstract supertype for all nodes.

An AbstractNode object is a point in space.

Common methods:

coordinates

dimension

dofs

ONSAS.Elements.Dof — Type

Degree of freedom struct. This is a scalar degree of freedom of the structure.

Fields:

index – degree of freedom identification number.

ONSAS.Elements.Node — Type

A Node is a point in space. The coordinates of the node are stored using SVector data type from static arrays.

Fields:

x – stores the coordinates.

dofs – stores the node degrees of freedom, maps symbol to dofs.

ONSAS.Elements.Node — Method

Node constructor with an AbstractVector v.

ONSAS.Elements.Node — Method

Node constructor with a NTuple t and an optional Dof dictionary dofs.

ONSAS.Elements.Node — Method

1D Node constructor with coordinates x₁.

ONSAS.Elements.Node — Method

3D Node constructor with coordinates x₁, x₂ and x₃ and an optional Dof dictionary dofs..

ONSAS.Elements.Node — Method

2D Node constructor with coordinates x₁ and x₂.

ONSAS.Elements.Tetrahedron — Type

A Tetrahedron represents a 3D volume element with four nodes.

Fields:

nodes – stores the tetrahedron nodes.

label – stores the tetrahedron label.

References

See [Belytschko].

ONSAS.Elements.Tetrahedron — Method

Constructor for a Tetrahedron element without nodes and a label. This function is used to create meshes via GMSH.

ONSAS.Elements.TriangularFace — Type

A TriangularFace represents an element composed by three Nodes.

Fields:

nodes – stores triangle nodes.

label – stores the triangle label.

ONSAS.Elements.TriangularFace — Method

Constructor for a TriangularFace element considering the nodes n₁ n₂ and n₃.

ONSAS.Elements.TriangularFace — Method

Constructor for a TriangularFace element without nodes and a label. This function is used to create meshes via GMSH.

ONSAS.Elements.Truss — Type

A Truss represents an element composed by two Nodes that transmits axial force only.

Fields:

nodes – stores the truss nodes.

cross_sections – stores the truss cross-section properties.

label – stores the truss label.

References

See [ANLE].

ONSAS.Elements.Truss — Method

Constructor for a Truss element without nodes and a label. This function is used to create meshes via GMSH.

ONSAS.Elements.Truss — Method

Constructor for a Truss element considering the nodes n₁ and n₂ and the cross-section g.

Base.:∈ — Method

Returns true if a point p is inside the AbstractElement e.

Base.:∈ — Method

Checks if a point p is inside a Tetrahedron element t.

Base.getindex — Method

Returns the coordinate at component i from the AbstractNode n.

Base.getindex — Method

Returns the entry of a Vector v at index corresponding to the Dof d index.

Base.getindex — Method

Returns a Vector of entries of a Vector v at indexes corresponding to the Dofs vector vd.

Base.maximum — Method

Returns the maximum dof index from a Vector of Dofs vd

Base.setindex! — Method

Sets a Vector of values tv to a Vector v at indexes corresponding to the Dofs vector vd.

Base.setindex! — Method

Sets the index of a Vector v with the index corresponding to the Dof d.

Dictionaries.index — Method

Returns the dof index of the Dof d

Dictionaries.index — Method

Returns the dof index from a Vector of Dofs vd

ONSAS.CrossSections.area — Method

Returns the AbstractFace f area.

ONSAS.CrossSections.area — Method

Returns the area of a TriangularFace element tf.

ONSAS.Elements._X_rows — Method

Returns auxiliar matrices with the element coordinates at the deformed and reference configurations.

ONSAS.Elements._area_vec — Method

Returns the area vector with direction and modulus of a TriangularFace element tf.

ONSAS.Elements._aux_b — Method

Returns auxiliar vectors bref and bdef of a truss element.

ONSAS.Elements._aux_matrices — Method

Returns

ONSAS.Elements._coordinates_matrix — Method

Returns the reshaped coordinates elem_coords of the tetrahedron element into a 4x3 matrix.

ONSAS.Elements._interpolation_matrix — Method

Returns the interpolation matrix 𝑀 for a Tetrahedron element t.

ONSAS.Elements._jacobian_mat — Method

Computes Jacobian matrix

ONSAS.Elements._lengths — Method

Returns deformed and reference lengths of a truss element.

ONSAS.Elements._shape_functions_derivatives — Function

Returns the shape functions derivatives of a Tetrahedron element.

ONSAS.Elements._strain — Method

Returns the rotated engineering strain for a given reference and deformed length l_ini and the deformed length l_def.

ONSAS.Elements._volume — Method

Computes volume element of a tetrahedron given J = det(𝔽).

ONSAS.Elements.add! — Method

Sets a Vectors of dofs vd to the AbstractNode n assigned to the symbol s.

ONSAS.Elements.coordinates — Method

Returns the AbstractEntity e coordinates.

ONSAS.Elements.coordinates — Method

Returns the AbstractNode n coordinates.

ONSAS.Elements.coordinates — Method

Returns each AbstractNode coordinates in a Vector of Nodes vn.

ONSAS.Elements.coordinates — Method

Returns each AbstractFace coordinates in a Vector of Faces vf.

ONSAS.Elements.create_entity — Method

Returns an AbstractEntity given an empty AbstractEntity e and a Vector of Nodes vn.

ONSAS.Elements.create_entity — Method

Returns a Tetrahedron given an empty Tetrahedron t and a Vector of Nodes vn.

ONSAS.Elements.create_entity — Method

Returns a TriangularFace given an empty TriangularFace tf and a Vector of Nodes vn.

ONSAS.Elements.create_entity — Method

Returns a Tetrahedron given an empty Tetrahedron t and a Vector of Nodes vn.

ONSAS.Elements.cross_section — Method

Returns the AbstractElement e cross_section.

ONSAS.Elements.cross_section — Method

Returns the cross-section of a Truss element t.

ONSAS.Elements.dimension — Method

Returns the AbstractEntity dimension.

ONSAS.Elements.dimension — Method

Returns the AbstractNode n dimension (1D, 2D or 3D).

ONSAS.Elements.dofs — Method

Returns the dofs of an AbstractEntity e.

ONSAS.Elements.dofs — Method

Returns AbstractNode n degrees of freedom with symbol s.

ONSAS.Elements.dofs — Method

Returns AbstractNode n degrees of freedom.

ONSAS.Elements.dofs — Method

Returns the dofs of a Vector ve with AbstractEntityes.

ONSAS.Elements.dofs — Method

Returns degrees of freedom for each AbstractNode in a vector of nodes vn.

ONSAS.Elements.inertial_forces — Method

Returns the inertial forces vector of an AbstractElement e. with an AbstractMaterial m

ONSAS.Elements.internal_forces — Method

Returns the internal forces vector of an AbstractElement e with an AbstractMaterial m.

ONSAS.Elements.internal_forces — Method

Returns the internal force of a Tetrahedron element t doted with an AbstractMaterial m + and a an element displacement vector u_e.

ONSAS.Elements.internal_forces — Method

Returns the internal force of a Truss element t formed by an AbstractMaterial m and a an element displacement vector u_e.

ONSAS.Elements.local_dof_symbol — Method

Returns local dofs symbols of the AbstractElement e (for linear displacements :u is used) in a vector. Since global degrees of freedom are for the assemble process this function is used to compute the global dofs of the element by extracting the node dofs with the symbol defined by the AbstractElement e.

ONSAS.Elements.local_dof_symbol — Method

Returns the local dof symbol of a Tetrahedron element.

ONSAS.Elements.local_dof_symbol — Method

Returns the local dof symbol of a Truss element.

ONSAS.Elements.local_dofs — Method

Returns local dofs given a vector of local dof symobls. This method extracts all node dofs with the same symbol as localdofsymbol

ONSAS.Elements.nodes — Method

Returns the Nodes an AbstractEntity e.

ONSAS.Elements.normal_direction — Method

Returns the AbstractFace f normal.

ONSAS.Elements.normal_direction — Method

Returns the normal direction n of a TriangularFace element tf.

ONSAS.Elements.strain — Method

Returns the AbstractElement e strain.

ONSAS.Elements.strain — Method

Returns the strain of given Truss element t with a element displacement vector u_e.

ONSAS.Elements.stress — Method

Returns the AbstractElement e stress.

ONSAS.Elements.stress — Method

Returns the stress of given Truss element t with a element displacement vector u_e.

ONSAS.Elements.volume — Method

Returns the Tetrahedron t volume in the reference configuration.

ONSAS.Elements.weights — Method

Returns the weights to interpolate a scalar field at the Nodes Dof corresponding to the AbstractElement e.

ONSAS.Elements.weights — Method

Returns the interpolation weights of a point p in a Tetrahedron element t.

ONSAS.Utils.label — Method

Returns the label of an AbstractEntity e.