Meshes Interface

This section describes the interfaces for different Meshes types.

ONSAS.Meshes — Module

Module defining meshes entities interface. Each mesh consists of a data type with the nodes and elements. Moreover, different sets of nodes and elements can be defined.

ONSAS.Meshes.AbstractMesh — Type

Abstract supertype for all meshes.

The following methods are provided by the interface:

Common methods:

dimension

dofs

num_dofs

elements

faces

num_elements

nodes

num_nodes

ONSAS.Meshes.Mesh — Type

A Mesh is a collection of Elements, Faces and Nodes that cover the discretized domain, together with Sets of elements and nodes.

Methods:

node_set

add_node_to_set!

element_set

add_element_to_set!

face_set

add_face_to_set!

ONSAS.Meshes.Mesh — Method

Constructor for Mesh with a Node'sVector and extra.

Base.push! — Method

Push a new Element into the AbstractMesh m and return the new Element position.

Base.push! — Method

Push a new Face f into the AbstractMesh m and return the new Face position.

Base.push! — Method

Push a new Node into the AbstractMesh m and return the new Node position.

Base.push! — Method

Push a new vector of Elements into the AbstractMesh m.

Base.push! — Method

Push a new Vector of Faces vf into the AbstractMesh m.

Base.push! — Method

Push a new Vector ofNodes into theAbstractMeshm`.

ONSAS.Elements.dimension — Method

Return the dimension of an AbstractMesh.

ONSAS.Elements.nodes — Method

Return a Vector of Nodes defined in the AbstractMesh m.

ONSAS.Elements.nodes — Method

Return a Vector of Nodes with node_set_name in the Mesh m.

ONSAS.Elements.num_nodes — Method

Return the number of Nodes of the AbstractMesh m.

ONSAS.Meshes._isempty_dofs — Method

Return true if the AbstractMesh m has Dofs defined.

ONSAS.Meshes.add_element_to_set! — Method

Add an Element e to the Set element_set_name.

ONSAS.Meshes.add_element_to_set! — Method

Add element_id to the Mesh m Set element_set_name.

ONSAS.Meshes.add_face_to_set! — Method

Add a Face n to the Set face_set_name.

ONSAS.Meshes.add_face_to_set! — Method

Add face_id to the Mesh m Set face_set_name.

ONSAS.Meshes.add_node_to_set! — Method

Add a Node n to the Set node_set_name.

ONSAS.Meshes.add_node_to_set! — Method

Add a node_id to the Mesh m Set node_set_name.

ONSAS.Meshes.element — Method

Return the Elements of the AbstractMesh m at index i.

ONSAS.Meshes.element_set — Method

Return the Mesh m Element Sets.

ONSAS.Meshes.element_set — Method

Return the Mesh m Element Set with element_set_name.

ONSAS.Meshes.elements — Method

Return the Elements of the AbstractMesh m.

ONSAS.Meshes.elements — Method

Return a Vector of Elements with element_set_name in the Mesh m.

ONSAS.Meshes.face_set — Method

Return the Mesh m Face Sets.

ONSAS.Meshes.face_set — Method

Return the Mesh m Element Set with element_set_name.

ONSAS.Meshes.faces — Method

Return a Vector of Faces defined in the AbstractMesh m.

ONSAS.Meshes.faces — Method

Return a Vector of Faces with face_set_name in the Mesh m.

ONSAS.Meshes.node_set — Method

Return the Mesh m Node Sets.

ONSAS.Meshes.node_set — Method

Return the Mesh m Node Set with node_set_name.

ONSAS.Meshes.num_dofs — Method

Return the number of Dofs defined in the AbstractMesh m. This function assumes that Dofs indexes start from Dof(1)

ONSAS.Meshes.num_elements — Method

Return the number of elements of the AbstractMesh m.

ONSAS.Meshes.num_faces — Method

Return the number of Faces of the AbstractMesh m.

ONSAS.Utils.apply! — Method

Adds n dofs_per_node Dofs with dof_symbol to the AbstractMesh m.

ONSAS.Utils.dofs — Method

Return the AbstractMesh vector of Dofs. Entry i contains the Dofs of node with index i in the AbstractMesh vector of nodes.

Modules = [ONSAS.Meshes.Gmsh]