Meshes Interface

This section describes the interfaces for different Meshes types.

ONSAS.Meshes — Module

Module defining meshes entities interface. Each mesh consists of a data type with the nodes and elements. Moreover, different sets of nodes and elements can be defined.

ONSAS.Meshes.AbstractMesh — Type

Abstract supertype for all meshes.

The following methods are provided by the interface:

Common methods:

dimension

dofs

num_dofs

elements

faces

num_elements

nodes

num_nodes

ONSAS.Meshes.Mesh — Type

Mesh. A Mesh is a collection of Elements, Faces and Nodes that cover the discretized domain, together with Sets of elements and nodes.

Fields:

nodes – stores the dim dimensional Nodes of the grid.

faces – stores the Faces of the grid.

elements – stores the Elements of the mesh.

ONSAS.Meshes.PointEvalHandler — Type

PointEvalHandler struct. A PointEvalHandler facilitates the process of evaluating a solution at a given vector of points obtained at the Nodes Dofs in a Mesh.

Fields:

vec_points – stores the points where the solution will be evaluated.

mesh – stores the mesh where the solution is obtained.

interpolator – stores the interpolator used to evaluate the solution.

ONSAS.Meshes.PointEvalHandler — Method

Constructor of a PointEvalHandler from a Mesh and a AbstractVector of Points .

ONSAS.Meshes.PointEvalHandler — Method

Constructor of a PointEvalHandler from a Mesh and a Point .

ONSAS.Meshes.PointsInterpolator — Type

PointsInterpolator struct. A PointsInterpolator struct stores the weights nodes and elements needed to interpolate the solution at a given point. The index of each Vector is the index in the Vector of Points.

Fields:

node_to_weights – stores a Dictionary with Nodes as keys and the corresponding weights as values.

point_to_element – stores the Element where the point is located.

Base.push! — Method

Pushes a new Element into the AbstractMesh m.

Base.push! — Method

Pushes a new Face f into the AbstractMesh m.

Base.push! — Method

Pushes a new Node into the AbstractMesh m.

Base.push! — Method

Pushes a new vector of Elements into the AbstractMesh m.

Base.push! — Method

Pushes a new Vector of Faces vf into the AbstractMesh m.

Base.push! — Method

Pushes a new Vector ofNodes into theAbstractMeshm`.

ONSAS.Elements.add! — Method

Adds n dofs_per_node Dofs with dof_symbol to the AbstractMesh m.

ONSAS.Elements.dimension — Method

Returns the dimension of an AbstractMesh.

ONSAS.Elements.dofs — Method

Returns the AbstractMesh vector of Dofs. Entry i contains the Dofs of node with index i in the AbstractMesh vector of nodes.

ONSAS.Elements.nodes — Method

Returns a Vector of Nodes defined in the AbstractMesh m.

ONSAS.Meshes._isempty_dofs — Method

Returns true if the AbstractMesh m has Dofs defined.

ONSAS.Meshes.elements — Method

Returns the Elements of the AbstractMesh m.

ONSAS.Meshes.faces — Method

Returns a Vector of Faces defined in the AbstractMesh m.

ONSAS.Meshes.interpolator — Method

Returns the Interpolator used to evaluate the solution.

ONSAS.Meshes.mesh — Method

Returns the Mesh where the solution is obtained.

ONSAS.Meshes.node_to_weights — Method

Returns a Vector that maps the weight corresponding to each Node.

ONSAS.Meshes.num_dofs — Method

Returns the number of Dofs defined in the AbstractMesh m. This function assumes that Dofs indexes start from Dof(1)

ONSAS.Meshes.num_elements — Method

Returns the number of elements of the AbstractMesh m.

ONSAS.Meshes.num_nodes — Method

Returns the number of Nodes of the AbstractMesh m.

ONSAS.Meshes.points — Method

Returns the Vector of Points where the solution will be evaluated.

ONSAS.Meshes.points_to_element — Method

Returns a Vector that maps the Element where each Point is located.

ONSAS.Meshes.Gmsh — Module

Module defining meshes created in GMSH software.

ONSAS.Meshes.Gmsh.MshFile — Type

MshFile. A MshFile is a collection of Nodes. Also the filename

Fields:

filename – stores the .geo file name.

vec_nodes – stores the Nodess of the mesh.

connectivity – stores a Vector of Vector with the node indexes.

material_labels – stores a Vector of Strings with the material type labels defined in the .geo.

element_face_labels – stores a Vector of Strings with the face and element type labels defined in the .geo.

bc_labels – stores a Vector of Strings with the boundary condition type labels defined in the .geo.

ONSAS.Meshes.Gmsh.MshFile — Method

Constructor of the MshFile object with a file name filename.

ONSAS.Elements.dimension — Method

Returns the mesh dimension defined in the MshFile mfile.

ONSAS.Elements.nodes — Method

Returns the Nodess of the MshFile mfile.

ONSAS.Meshes.Gmsh._getlabels — Method

Returns material, entities and boundary conditions labels defined in physical_names.

ONSAS.Meshes.Gmsh.bc_label — Method

Returns boundary conditions label for the AbstractEntity with index entity_index.

ONSAS.Meshes.Gmsh.bc_label — Method

Returns boundary conditions labels defined in the MshFilemfile.

ONSAS.Meshes.Gmsh.connectivity — Method

Returns the connectivity defined in the MshFile mfile.

ONSAS.Meshes.Gmsh.entity_label — Method

Returns entities label for the AbstractEntity with index entity_index.

ONSAS.Meshes.Gmsh.entity_label — Method

Returns entities labels defined in the MshFilemfile.

ONSAS.Meshes.Gmsh.material_label — Method

Returns material label for the AbstractEntity with index entity_index.

ONSAS.Meshes.Gmsh.material_label — Method

Returns material labels defined in the MshFilemfile.

ONSAS.Meshes.Gmsh.physical_index — Method

Returns physical_index defined in the MshFile mfile.

ONSAS.Meshes.Gmsh.physical_index — Method

Returns physical indexes defined in the MshFile mfile.