Static Analysis Interface

This section describes the interfaces to handle with static analysis.

ONSAS.StructuralAnalyses.StaticAnalyses.StaticAnalysis — Type

Constructor for StaticAnalysis given a final time (or load factor) t₁ and the number of steps NSTEPS.

ONSAS.StructuralAnalyses.StaticAnalyses.StaticAnalysis — Type

StaticAnalysis struct. A StaticAnalysis is a collection of parameters for defining the static analysis of the structure. In the static analysis, the structure is analyzed at a given load factor (this variable is analog to time).

Fields:

s – Stores the structure to be analyzed.

state – Stores the structural state.

λᵥ – Stores the load factors vector of the analysis

current_step – Stores the current load factor step

ONSAS.StructuralAnalyses.StaticAnalyses.StaticState — Type

An StaticState object facilitates the process of storing the relevant static variables of the structure during the displacements iteration.

Fields:

s – stores the structure

ΔUᵏ – stores displacements vector increment.

Uᵏ – stores displacements vector.

Fₑₓₜᵏ – stores external forces vector.

Fᵢₙₜᵏ – stores internal forces vector.

Kₛᵏ – stores the system tangent matrix.

ϵᵏ – stores a vector with strains for each element.

σᵏ – stores a vector with stresses for each element.

assembler – assembler handler object

iter_state – current Δu iteration state

ONSAS.StructuralAnalyses.StaticAnalyses.StaticState — Method

Constructor for StaticState given an AbstractStructure s.

Base.push! — Method

Pushes the current state c_state into the StatesSolution st_sol.

ONSAS.StructuralAnalyses.StaticAnalyses.current_load_factor — Method

Returns the current load factor of an StaticAnalysis sa.

ONSAS.StructuralAnalyses.StaticAnalyses.load_factors — Method

Returns the final load factor vector of an StaticAnalysis sa.

ONSAS.StructuralAnalyses._next! — Method

Jumps to the next current load factor defined in the StaticAnalysis sa.

ONSAS.StructuralAnalyses.current_time — Method

Returns the current load factor of an StaticAnalysis sa.

ONSAS.StructuralAnalyses.final_time — Method

Returns the final load factor of an StaticAnalysis sa.

ONSAS.StructuralAnalyses.initial_time — Method

Returns the initial load factor of an StaticAnalysis sa.

ONSAS.StructuralAnalyses.is_done — Method

Returns true if the StaticAnalysis sa is completed.

ONSAS.StructuralAnalyses.reset! — Method

Rests the current load factor of the StaticAnalysis sa (set to 1).

ONSAS.StructuralAnalyses.residual_forces — Method

Returns the current residual forces of the StaticState sc.

ONSAS.StructuralAnalyses.tangent_matrix — Method

Returns the current system tangent matrix of the StaticState sc.

ONSAS.StructuralSolvers._assemble! — Method

Assembles the Structure s (internal forces) during the StaticAnalysis sa.

ONSAS.StructuralSolvers._reset! — Method

Resets the StaticState assembled magnitudes.

ONSAS.StructuralSolvers._solve — Method

Solves an StaticAnalysis sa with an AbstractSolver alg.

ONSAS.StructuralSolvers._step! — Method

Computes ΔU for solving the StaticAnalyses sa with a NewtonRaphson method.

ONSAS.StructuralSolvers._update! — Method

Updates displacements in the StaticState sc with a displacements increment vector ΔU.