StaticAnalyses

ONSAS.StaticAnalyses — Module

Module defining static analyses. Each static state analysis is defined by an structure and a load factors vector. During the analysis the static state of the structure is mutated through the load factors vector.

ONSAS.StaticAnalyses.AbstractStaticAnalysis — Type

Abstract supertype for all structural analysis.

An AbstractStaticAnalysis object facilitates the process of defining an static analysis to be solved. Time variable for static analysis is used to obtain a load factor value. Of course this abstract type inherits from AbstractStructuralAnalysis type, and extends the following methods:

Abstract Methods

initial_time

current_time

final_time

load_factors

current_load_factor

next!

is_done

reset!

_solve!

step!

Common fields

λᵥ – stores the load factors vector.

current_step – stores the current step of the analysis.

state – stores the current state of the analysis.

Base.push! — Method

Stores the current state into the solution.

ONSAS.Assemblers.assemble! — Method

Assembles the Structure s (internal forces) during the StaticAnalysis sa.

ONSAS.Assemblers.reset! — Method

Sets the current load factor of the structural analysis to the initial load factor. Also Reset! the iteration and AbstractStructuralState.

ONSAS.StaticAnalyses.current_load_factor — Method

Return the current load factor of an structural analysis.

ONSAS.StaticAnalyses.load_factors — Method

Return the final load factor vector of an structural analysis.

ONSAS.StaticAnalyses.reset_assemble! — Method

Reset the assembled magnitudes in the state.

ONSAS.StructuralAnalyses.current_time — Method

Return the current load factor of an structural analysis.

ONSAS.StructuralAnalyses.final_time — Method

Return the final load factor of an structural analysis.

ONSAS.StructuralAnalyses.initial_time — Method

Return the initial load factor of an structural analysis.

ONSAS.StructuralAnalyses.is_done — Method

Return true if the structural analysis is completed.

ONSAS.StructuralSolvers.next! — Method

Jumps to the next current load factor defined in the structural analysis.