StructuralAnalyses

ONSAS.StructuralAnalyses — Module

Module defining structural analyses that can be solved. Each structural analysis consists of a data type with an structure to be analyzed, analysis parameters and an specific state which describes the current state of structure in the analysis.

ONSAS.StructuralAnalyses.AbstractStaticState — Type

States representing static analyses.

ONSAS.StructuralAnalyses.AbstractStructuralAnalysis — Type

Abstract supertype for all structural analysis.

An structural analysis object facilitates the process of defining an structural analysis to be solved.

Abstract Methods

structure

free_dofs

initial_time

current_time

final_time

current_state

current_iteration

next!

is_done

ONSAS.StructuralAnalyses.AbstractStructuralState — Type

Abstract supertype to define a new structural state. Abstract Methods

Accessors:

displacements

Δ_displacements

Δ_displacements!

external_forces

internal_forces

residual_forces!

tangent_matrix

strain

stress

structure

free_dofs

Iteration:

assemble!

assembler

iteration_residuals

residual_forces_norms

residual_displacements_norms

ONSAS.Assemblers.assemble! — Method

Assembles the element e stiffness matrix matrix K_e into the structural state.

ONSAS.Assemblers.assemble! — Method

Assembles the element e internal forces fᵢₙₜ_e into the structural state.

ONSAS.Assemblers.assemble! — Method

Assembles the element e stress σₑ and strain ϵₑ into the structural state.

ONSAS.Assemblers.end_assemble! — Method

Fill the system tangent matrix in the structural state once the assembler object is built.

ONSAS.Entities.internal_forces — Method

Return the current internal forces vector in the structural state.

ONSAS.Entities.strain — Method

Return strains for each Element in the structural state.

ONSAS.Entities.stress — Method

Return stresses for each Element in the structural state.

ONSAS.StructuralAnalyses.assembler — Method

Return the assembler used in the structural state.

ONSAS.StructuralAnalyses.current_iteration — Method

Return the current displacements iteration state of the structural analysis.

ONSAS.StructuralAnalyses.current_state — Method

Return the current state of the structural analysis.

ONSAS.StructuralAnalyses.current_time — Method

Return the current time of structural analysis.

ONSAS.StructuralAnalyses.displacements — Method

Return current displacements vector at the current structural state.

ONSAS.StructuralAnalyses.external_forces — Method

Return external forces vector in the structural state.

ONSAS.StructuralAnalyses.final_time — Method

Return the final time of structural analysis.

ONSAS.StructuralAnalyses.initial_time — Method

Return the initial time of structural analysis.

ONSAS.StructuralAnalyses.internal_cache — Method

Return the cache associated to the given element.

ONSAS.StructuralAnalyses.is_done — Method

Return true if the structural analysis is completed.

ONSAS.StructuralAnalyses.iteration_residuals — Method

Return current ResidualsIterationStep object form an structural state.

ONSAS.StructuralAnalyses.residual_displacements_norms — Method

Return relative residual displacements for the current structural state.

ONSAS.StructuralAnalyses.residual_forces! — Method

Return residual forces vector in the structural state.

ONSAS.StructuralAnalyses.residual_forces_norms — Method

Return relative residual forces for the current structural state.

ONSAS.StructuralAnalyses.structure — Method

Return analyzed structure in the structural analysis.

ONSAS.StructuralAnalyses.tangent_matrix — Method

Return system tangent matrix in the structural state given a solver.

ONSAS.StructuralAnalyses.Δ_displacements! — Method

Update and return current displacements increment vector at the current structural state.

ONSAS.StructuralAnalyses.Δ_displacements — Method

Return current displacements increment vector at the current structural state.

ONSAS.Structures.free_dofs — Method

Return free Dofs of the structure in the structural state.

ONSAS.Utils.apply! — Method

Apply a boundary condition to the structural analysis at the current analysis time.

ONSAS.Utils.apply! — Method

Apply a vector of load boundary conditions to the structure.