Structural Model Interface

This section describes the interfaces for defining structural models.

ONSAS.StructuralModel — Module

Module defining structure entities interface. Each Structure consists of a data type with materials, boundary conditions, mesh and a vector of free dofs.

ONSAS.Handlers.PointEvalHandler — Method

Constructor of a PointEvalHandler from a Structure and a AbstractVector of Points .

ONSAS.Meshes.Mesh — Method

Return the Mesh given an MshFile msh_file and a StructuralEntities s_entities.

ONSAS.StructuralModel.AbstractStructure — Type

Abstract supertype to define a new structure. An AbstractStructure object facilitates the process of assigning materials, elements, and boundary conditions to the mesh. Moreover this struct is used to solve an specific structural analysis.

Common methods:

Mesh:

dofs

num_dofs

free_dofs

num_free_dofs

nodes

num_nodes

faces

num_faces

elements

num_elements

mesh

Boundary conditions:

boundary_conditions

displacement_bcs

load_bcs

node_bcs

element_bcs

Materials:

materials

ONSAS.StructuralModel.StructuralBoundaryConditions — Type

Structural boundary conditions. A StructuralBoundaryConditions is a collection of BoundaryConditions defining the boundary conditions of the structure.

Fields:

node_bcs – Maps each boundary conditions for a vector of nodes.

face_bcs – Maps each boundary conditions for a vector of faces.

element_bcs – Maps each boundary conditions for a vector of elements.

ONSAS.StructuralModel.StructuralBoundaryConditions — Method

Constructor for empty StructuralBoundaryConditions with a Vector of AbstractBoundaryConditions vbc.

ONSAS.StructuralModel.StructuralEntities — Type

Structural elements struct. A StructuralMaterials is a collection of Elements and Facess types assigning to a vector of Faces and Elements. This struct is used to define meshes via GMSH.

Fields:

elem_types_to_elements – Store a dictionary with Element types (Elements without assigned Nodes)

as keys and the corresponding elements created.

face_types_to_faces – Store a dictionary with Face types (Faces without assigned Nodes)

as keys and the corresponding faces created.

ONSAS.StructuralModel.StructuralEntities — Method

Constructor for an empty StructuralEntities with a Vector of Elements velems and Faces vfaces.

ONSAS.StructuralModel.StructuralMaterials — Type

A StructuralMaterials is a collection of Materials and Elements assigning materials to a vector of elements.

ONSAS.StructuralModel.StructuralMaterials — Method

Constructor for empty StructuralMaterials with a Vector of materials vmats. This will assign an empty Vector of Elements to each material.

ONSAS.StructuralModel.Structure — Type

Constructor of a Structure given a MshFile msh_file, StructuralMaterials materials, StructuralBoundaryConditions bcs.

ONSAS.StructuralModel.Structure — Type

An Structure object facilitates the process of assembling and creating the structural analysis.

ONSAS.StructuralModel.Structure — Method

Constructor with StructuralMaterials materials, StructuralBoundaryConditions bcs and AbstractMesh mesh seting fixed dofs with FixedDofBoundaryCondition defined in bcs

Base.delete! — Method

Delete the Material m from the StructuralMaterials sm.

Base.getindex — Method

Return the Vector of BoundaryConditionss applied to the AbstractElement e.

Base.getindex — Method

Return the Vector of BoundaryConditionss applied to the AbstractFace f.

Base.getindex — Method

Return the Vector of BoundaryConditionss applied to the AbstractNode n.

Base.getindex — Method

Return the Vector of Nodes and Elements where the BoundaryCondition bc is imposed.

Base.getindex — Method

Return the Vector of Material of the element e.

Base.getindex — Method

Return a the BoundaryConditions with the label l in the StructuralEntities sb.

Base.getindex — Method

Return the BoundaryCondition with the label l in the StructuralBoundaryConditions sb.

Base.getindex — Method

Return the Entity with the label l in the StructuralEntities s_entities.

Base.getindex — Method

Return the Material mapped with the label l.

Base.getindex — Method

Return the Vector of Elements that are conformed by the Materialm`.

Base.insert! — Method

Insert the Material m into the StructuralMaterials sm.

Base.pairs — Method

Return Pairs of Material and Element in the StructuralMaterials sm.

Base.push! — Method

Pushes to the AbstractElement e the BoundaryCondition bc in the StructuralBoundaryConditions sb.

Base.push! — Method

Pushes to the AbstractFace f the BoundaryCondition bc in the StructuralBoundaryConditions sb.

Base.push! — Method

Pushes to the AbstractNode n the BoundaryCondition bc in the StructuralBoundaryConditions sb.

Base.replace! — Function

Replace the AbstractMaterial with the label l for the new material new_m in the Structure. The previous material Elements are assigned to the new.

Base.replace! — Function

Replace the AbstractMaterial with the label l for the new material new_m in the StructuralMaterials sm. The previous material Elements are assigned to the new.

ONSAS.BoundaryConditions._apply — Method

Return a Vector of Dofs and values to apply a LoadDofBoundaryCondition and a set of StructuralBoundaryConditions bcs at time t.

ONSAS.BoundaryConditions._apply — Method

Return a Vector of Dofs to delete given a FixedDofBoundaryCondition and a set of StructuralBoundaryConditions bcs.

ONSAS.BoundaryConditions._apply — Method

Return a Vector of FixedDofBoundaryCondition f_bcs and a set of StructuralBoundaryConditions bcs.

ONSAS.Elements.apply! — Method

Apply the StructuralBoundaryConditions to the AbstractMesh m. For this is required sets into the Mesh and the corresponding boundary condition labels declared in bcs.

ONSAS.Elements.apply! — Method

Apply the StructuralBoundaryConditions to the AbstractMesh m. For this is required sets into the Mesh and the corresponding boundary condition labels declared in bcs.

ONSAS.Elements.dofs — Method

Return the Dofs of the AbstractStructure s

ONSAS.Elements.nodes — Method

Return a Vector of Nodes defined in the AbstractStructure s.

ONSAS.Elements.num_nodes — Method

Return the number of Nodes of the AbstractStructure s

ONSAS.Handlers.mesh — Method

Return the Mesh of the AbstractStructure s

ONSAS.Meshes.elements — Method

Return the Elements of the AbstractStructure s

ONSAS.Meshes.faces — Method

Return the Faces of the AbstractStructure s

ONSAS.Meshes.num_dofs — Method

Return the number of Dofs of the AbstractStructure s

ONSAS.Meshes.num_elements — Method

Return the number of Elements of the AbstractStructure s

ONSAS.Meshes.num_faces — Method

Return the number of Faces of the AbstractStructure s

ONSAS.StructuralModel._add_entity_to_set! — Method

Add an entity index of type AbstractElement to the Mesh m element Sets.

ONSAS.StructuralModel._add_entity_to_set! — Method

Add an entity index of type AbstractFace to the Mesh m face Sets.

ONSAS.StructuralModel._add_entity_to_set! — Method

Add an entity index of type AbstractNode to the Mesh m node Sets.

ONSAS.StructuralModel._apply_element_bcs! — Method

Apply Element boundary conditions given the AbstractMesh m to the StructuralBoundaryConditions bcs.

ONSAS.StructuralModel._apply_face_bcs! — Method

Apply Face boundary conditions given the AbstractMesh m to the StructuralBoundaryConditions bcs.

ONSAS.StructuralModel._apply_node_bcs! — Method

Apply Node boundary conditions given the AbstractMesh m to the StructuralBoundaryConditions bcs.

ONSAS.StructuralModel._delete_empty_bcs! — Method

Delete empty boundary conditions from the StructuralBoundaryConditions bcs.

ONSAS.StructuralModel._element_material_is_unique — Method

Checks that each Element has a single Material in the dictionary mat_dict.

ONSAS.StructuralModel.all_bcs — Method

Return all BoundaryConditionss defined into StructuralBoundaryConditions.

ONSAS.StructuralModel.all_entities — Method

Return all Entitys defined into StructuralEntities.

ONSAS.StructuralModel.boundary_conditions — Method

Return the StructuralBoundaryConditions of the AbstractStructure s

ONSAS.StructuralModel.displacement_bcs — Method

Return the DisplacementBoundaryConditions of the AbstractStructure s

ONSAS.StructuralModel.displacement_bcs — Method

Return a Vector of DisplacementBoundaryConditions applied to Nodes and Elements in the StructuralBoundaryConditions se.

ONSAS.StructuralModel.elem_types — Method

Return the Vector of Element types defined in the StructuralEntities s_entities.

ONSAS.StructuralModel.elem_types_to_elements — Method

Return a Dictionary with Element types as keys and the corresponding Elements as values.

ONSAS.StructuralModel.element_bcs — Method

Return the AbstractBoundaryConditions imposed to Elements in the AbstractStructure s

ONSAS.StructuralModel.element_bcs — Method

Return the Dictionary of BoundaryConditionss applied to Elements.

ONSAS.StructuralModel.element_materials — Method

Return a Dictionary with Materials as keys and the corresponding Elements as values.

ONSAS.StructuralModel.face_bcs — Method

Return the dictionary of BoundaryConditionss applied to Faces.

ONSAS.StructuralModel.face_types — Method

Return the Vector of Face types defined in the StructuralEntities s_entities.

ONSAS.StructuralModel.face_types_to_faces — Method

Return a Dictionary with Face types as keys and the corresponding Faces as values.

ONSAS.StructuralModel.fixed_dof_bcs — Method

Return a Vector of FixedDofBoundaryConditions applied to Nodes, Faces and Elements. in the StructuralBoundaryConditions se.

ONSAS.StructuralModel.free_dofs — Method

Return free Dofs of the structure s

ONSAS.StructuralModel.load_bcs — Method

Return the LocalPressureBoundaryConditions of the AbstractStructure s

ONSAS.StructuralModel.load_bcs — Method

Return a Vector of FixedDofBoundaryConditions applied to Nodes and Elements in the StructuralBoundaryConditions bcs.

ONSAS.StructuralModel.materials — Method

Return the StructuralMaterialss of the AbstractStructure s

ONSAS.StructuralModel.node_bcs — Method

Return the AbstractBoundaryConditions imposed to Nodes in the AbstractStructure s

ONSAS.StructuralModel.node_bcs — Method

Return the dictionary of BoundaryConditionss applied to Nodes.

ONSAS.StructuralModel.num_free_dofs — Method

Return the number of free Dofs of the AbstractStructure s