StructuralBoundaryConditions

ONSAS.StructuralBoundaryConditions — Module

Module defining the structural boundary conditions assigned to each entitiy. This data type includes:

AbstractBoundaryConditions assigned to Nodes

AbstractBoundaryConditions assigned to Faces

AbstractBoundaryConditions assigned to Elements

ONSAS.StructuralBoundaryConditions.StructuralBoundaryCondition — Type

Structural boundary conditions. A StructuralBoundaryCondition is a collection of BoundaryConditions defining the boundary conditions of the structure.

ONSAS.StructuralBoundaryConditions.StructuralBoundaryCondition — Method

Constructor for empty StructuralBoundaryCondition with a Vector of AbstractBoundaryConditions vbc.

Base.getindex — Method

Return the Vector of BoundaryConditionss applied to the AbstractElement e.

Base.getindex — Method

Return the Vector of BoundaryConditionss applied to the AbstractFace f.

Base.getindex — Method

Return the Vector of BoundaryConditionss applied to the AbstractNode n.

Base.getindex — Method

Return the BoundaryCondition with the label l in the StructuralBoundaryCondition sb.

Base.getindex — Method

Return the Vector of Nodes and Elements where the BoundaryCondition bc is imposed.

Base.getindex — Method

Return a the BoundaryConditions with the label l in the StructuralEntity sb.

Base.push! — Method

Pushes to the AbstractElement e the BoundaryCondition bc in the StructuralBoundaryCondition sb.

Base.push! — Method

Pushes to the AbstractFace f the BoundaryCondition bc in the StructuralBoundaryCondition sb.

Base.push! — Method

Pushes to the AbstractNode n the BoundaryCondition bc in the StructuralBoundaryCondition sb.

Base.replace! — Function

Replace the bc with the a label for a bc in the structural boundary conditions. The previous boundary conditions entities are assigned to the new.

ONSAS.BoundaryConditions.apply — Method

Return a Vector of Dofs and values to apply a LoadDofBoundaryCondition and a set of StructuralBoundaryCondition bcs at time t.

ONSAS.BoundaryConditions.apply — Method

Return a Vector of Dofs to delete given a FixedDof and a set of StructuralBoundaryCondition bcs.

ONSAS.BoundaryConditions.apply — Method

Return a Vector of FixedDof f_bcs and a set of StructuralBoundaryCondition bcs.

ONSAS.StructuralBoundaryConditions.all_bcs — Method

Return all BoundaryConditionss defined into StructuralBoundaryCondition.

ONSAS.StructuralBoundaryConditions.apply_element_bcs! — Method

Apply Element boundary conditions given the AbstractMesh m to the StructuralBoundaryCondition bcs.

ONSAS.StructuralBoundaryConditions.apply_face_bcs! — Method

Apply Face boundary conditions given the AbstractMesh m to the StructuralBoundaryCondition bcs.

ONSAS.StructuralBoundaryConditions.apply_node_bcs! — Method

Apply Node boundary conditions given the AbstractMesh m to the StructuralBoundaryCondition bcs.

ONSAS.StructuralBoundaryConditions.displacement_bcs — Method

Return a Vector of AbstractDisplacementBoundaryConditions applied to Nodes and Elements in the StructuralBoundaryCondition se.

ONSAS.StructuralBoundaryConditions.element_bcs — Method

Return the Dictionary of BoundaryConditionss applied to Elements.

ONSAS.StructuralBoundaryConditions.face_bcs — Method

Return the dictionary of BoundaryConditionss applied to Faces.

ONSAS.StructuralBoundaryConditions.fixed_dof_bcs — Method

Return a Vector of FixedDofs applied to Nodes, Faces and Elements. in the StructuralBoundaryCondition se.

ONSAS.StructuralBoundaryConditions.load_bcs — Method

Return a Vector of FixedDofs applied to Nodes and Elements in the StructuralBoundaryCondition bcs.

ONSAS.StructuralBoundaryConditions.node_bcs — Method

Return the dictionary of BoundaryConditionss applied to Nodes.

ONSAS.Utils.apply! — Method

Apply the StructuralBoundaryCondition to the AbstractMesh m. For this is required sets into the Mesh and the corresponding boundary condition labels declared in bcs.