Structural Solvers Interface

This section describes the interfaces for structural solvers.

ONSAS.StructuralSolvers — Module

Module defining structural solvers that can be used to solved different analyses. Each solver consists of a data type with a convergence criterion and a the iteration status. A _step! method is used to perform a single iteration step.

ONSAS.StructuralSolvers.AbstractConvergenceCriterion — Type

Abstract supertype for all convergence criterion.

ONSAS.StructuralSolvers.AbstractSolver — Type

Abstract supertype for all direct integration methods.

ONSAS.StructuralSolvers.ConvergenceSettings — Type

Facilitates the process of defining and checking numerical convergence.

ONSAS.StructuralSolvers.MaxIterCriterion — Type

MaxIterCriterion criteria indicates that the maximum number of iterations has been reached.

ONSAS.StructuralSolvers.NotConvergedYet — Type

NotConvergedYet indicates that the current iteration has not converged.

ONSAS.StructuralSolvers.ResidualForceCriterion — Type

ResidualForceCriterion convergence criterion.

ONSAS.StructuralSolvers.ResidualsIterationStep — Type

Stores the convergence information of at the current iteration step.

ONSAS.StructuralSolvers.ΔUCriterion — Type

ΔUCriterion indicates displacements increment convergence criterion.

ONSAS.StructuralSolvers.ΔU_and_ResidualForce_Criteria — Type

ΔU_and_ResidualForce_Criteria convergence criterion indicates that both ΔU and residual forces converged .

ONSAS.StructuralSolvers._init — Method

Return the initialized analysis. By default, it Return the same analysis.

ONSAS.StructuralSolvers._reset! — Method

Sets the ResidualsIterationStep i_step iteration step with nothing.

ONSAS.StructuralSolvers._reset! — Method

Sets the ResidualsIterationStep i_step iteration step to 0.

ONSAS.StructuralSolvers._solve! — Method

Internal solve function to be overloaded by each analysis.

ONSAS.StructuralSolvers._step! — Method

Increments a ResidualsIterationStep i_step by 1.

ONSAS.StructuralSolvers._step! — Method

Computes a step in time on the analysis considering the numerical AbstractSolver solver.

ONSAS.StructuralSolvers._update! — Method

Updates the ResidualsIterationStep i_step current convergence criterion.

ONSAS.StructuralSolvers._update! — Method

Updates the ResidualsIterationStep i_step iteration step with the current values of the displacement residuals ΔU_norm and ΔU_rel, the norm of the residual forces increment Δr_norm, and the relative Δr_rel.

ONSAS.StructuralSolvers.criterion — Method

Return the current criterion of the ResidualsIterationStep i_step.

ONSAS.StructuralSolvers.displacement_tol — Method

Return displacements tolerance set in the ConvergenceSettings tol.

ONSAS.StructuralSolvers.displacement_tol — Method

Return the current absolute and relative residual forces norm of the ResidualsIterationStep i_step.

ONSAS.StructuralSolvers.isconverged! — Method

Updates the ResidualsIterationStep i_step given a ConvergenceSettings cs.

ONSAS.StructuralSolvers.iter — Method

Return the current iteration number of the ResidualsIterationStep i_step.

ONSAS.StructuralSolvers.max_iter_tol — Method

Return the maximum number of iterations set in the ConvergenceSettings tol.

ONSAS.StructuralSolvers.reset! — Function

Resets the analysis to the state before starting a new assembly.

ONSAS.StructuralSolvers.residual_forces_tol — Method

Return residual forces tolerance set in the ConvergenceSettings tol.

ONSAS.StructuralSolvers.residual_forces_tol — Method

Return the current absolute and relative residual forces norm of the ResidualsIterationStep i_step.

ONSAS.StructuralSolvers.solve! — Method

Solve a structural analysis problem with the given solver. This function mutates the state defined in the analysis problem; use solve to avoid mutation. For linear analysis problems, alg doesn't need to be provided.

Returns a solution structure holding the result and the algorithm used to obtain it.

ONSAS.StructuralSolvers.solve — Method

Solve a structural analysis problem with the given solver.

ONSAS.StructuralSolvers.step_size — Method

Return the step size set to the AbstractSolver solver .

ONSAS.StructuralSolvers.tolerances — Method

Return the numerical tolerances set to the AbstractSolver solver.