StructuralSolvers

ONSAS.StructuralSolvers — Module

Module defining structural solvers that can be used to solved different analyses. Each solver consists of a data type with a convergence criterion and a the iteration status. A step! method is used to perform a single iteration step.

ONSAS.StructuralSolvers.AbstractConvergenceCriterion — Type

Abstract supertype for all convergence criterion.

ONSAS.StructuralSolvers.AbstractSolver — Type

Abstract supertype for all direct integration methods.

ONSAS.StructuralSolvers.ConvergenceSettings — Type

Facilitates the process of defining and checking numerical convergence.

ONSAS.StructuralSolvers.DummySolver — Type

Solver used as placeholder for e.g. problems that don't require such struct.

ONSAS.StructuralSolvers.MaxIterCriterion — Type

MaxIterCriterion criteria indicates that the maximum number of iterations has been reached.

ONSAS.StructuralSolvers.NotConvergedYet — Type

NotConvergedYet indicates that the current iteration has not converged.

ONSAS.StructuralSolvers.ResidualForceCriterion — Type

ResidualForceCriterion convergence criterion.

ONSAS.StructuralSolvers.ResidualsIterationStep — Type

Stores the convergence information of at the current iteration step.

ONSAS.StructuralSolvers.ΔUCriterion — Type

ΔUCriterion indicates displacements increment convergence criterion.

ONSAS.StructuralSolvers.ΔU_and_ResidualForce_Criteria — Type

ΔU_and_ResidualForce_Criteria convergence criterion indicates that both ΔU and residual forces converged .

Base.show — Method

Show convergence settings.

CommonSolve.init — Method

Copy the structure and optionally reset the state.

CommonSolve.solve! — Function

Solve a structural analysis problem with the given solver, returning a solution structure which holds the result and the algorithm used to obtain it.

This function mutates the state defined in the analysis problem. Use solve to avoid mutation. For linear analysis problems, the algorithm doesn't need to be provided.

ONSAS.Assemblers.reset! — Method

Sets the iteration step with nothing.

ONSAS.Assemblers.reset! — Method

Sets the iteration step to 0.

ONSAS.StructuralAnalyses.tangent_matrix — Method

Return system tangent matrix in the structural state given a solver.

ONSAS.StructuralSolvers._solve! — Method

Internal solve function to be overloaded by each analysis.

ONSAS.StructuralSolvers.criterion — Method

Return the current convergence criterion.

ONSAS.StructuralSolvers.displacement_tol — Method

Return displacements tolerance set.

ONSAS.StructuralSolvers.displacement_tol — Method

Return the current absolute and relative residual forces norm.

ONSAS.StructuralSolvers.isconverged! — Method

Updates the convergence criteria.

ONSAS.StructuralSolvers.iterations — Method

Return the iterations done so far.

ONSAS.StructuralSolvers.max_iter_tol — Method

Return the maximum number of iterations set.

ONSAS.StructuralSolvers.next! — Method

Increment the time step given of a structural analysis. Dispatch is done for different solvers.

ONSAS.StructuralSolvers.residual_forces_tol — Method

Return residual forces tolerance set.

ONSAS.StructuralSolvers.residual_forces_tol — Method

Return the current absolute and relative residual forces norm.

ONSAS.StructuralSolvers.step! — Method

Increments a ResidualsIterationStep i_step by 1.

ONSAS.StructuralSolvers.step! — Method

Computes a step in time on the analysis considering the numerical AbstractSolver solver.

ONSAS.StructuralSolvers.step_size — Method

Return the step size.

ONSAS.StructuralSolvers.tolerances — Method

Return the numerical tolerances.

ONSAS.StructuralSolvers.update! — Method

Updates the current convergence criterion.

ONSAS.StructuralSolvers.update! — Method

Updates the iteration step with the current values of the displacement and forces residuals.